% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{add_layer_to_soil}
\alias{add_layer_to_soil}
\title{Split soil layer in two layers}
\usage{
add_layer_to_soil(x, il, w, method = c("interpolate", "exhaust"))
}
\arguments{
\item{x}{A numeric data.frame or matrix. Columns are soil layers.}

\item{il}{An integer value. The column/soil layer number after which a new layer is added.}

\item{w}{A numeric vector of length two. The weights used to calculate the values of the new layer.}

\item{method}{A character string. See \code{Details}.}
}
\value{
An object like x with one column more at position \code{il + 1}.
}
\description{
Split soil layer in two layers
}
\section{Details}{
 The method \code{interpolate} calculates the weighted mean of the
 columns/layers \code{il} and \code{il + 1}.
 The method \code{exhaust} distributes the value of \code{il + 1} according to the
 weights.
}

