% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Miscellaneous_Functions.R
\name{AdjMonthlyBioMass}
\alias{AdjMonthlyBioMass}
\title{Adjust mean monthly biomass values by climate input}
\usage{
AdjMonthlyBioMass(tr_VegBiom, do_adjBiom_by_temp = FALSE,
  do_adjBiom_by_ppt = FALSE, fgrass_c3c4ann, growing_limit_C = 4,
  isNorth = TRUE, MAP_mm = 450, monthly.temp)
}
\description{
Adjust mean monthly biomass values by climate input
}
\section{Default inputs}{

 - shrubs (IM_USC00107648_Reynolds; 70% shrubs, 30% C3): biomass was estimated at
   MAP = 450 mm/yr
 - sgs-grassland (GP_SGSLTER; 12% shrubs, 22% C3, and 66% C4): biomass was estimated at
   MAP = 340 mm/yr
}

