% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{adjust.WindspeedHeight}
\alias{adjust.WindspeedHeight}
\title{Function to extrapolate windspeeds measured at a height different from the
2-m above ground that are assumed by \pkg{SOILWAT2}}
\usage{
adjust.WindspeedHeight(uz, height)
}
\arguments{
\item{uz}{A numeric vector. Windspeed [m/s] at \code{height}.}

\item{height}{A numeric value. Height above ground at which \code{uz}
windspeed was measured.}
}
\value{
Windspeed [m/s] at a height of 2 m above ground.
}
\description{
Based on equation 33 in Allen et al. 2005. Note: "For wind measurements above
surfaces other than clipped grass, the user should apply the full logarithmic
equation B.14".
}
\references{
Allen R.G., Walter I.A., Elliott R., Howell T., Itenfisu D.,
  Jensen M. (2005) The ASCE standardized reference evapotranspiration
  equation. ASCE-EWRI Task Committee Report.
}
