% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soils_Functions.R
\name{adjustLayersDepth}
\alias{adjustLayersDepth}
\title{The wrapper only handles 1-cm resolution of soil depths
(mainly because of the \var{trco})}
\usage{
adjustLayersDepth(layers_depth, d)
}
\description{
The wrapper only handles 1-cm resolution of soil depths
(mainly because of the \var{trco})
}
