% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_ClimateDownscaling.R
\name{applyDeltas}
\alias{applyDeltas}
\title{Add/multiply deltas to historic daily data to generate future daily SOILWAT2-formatted weather.
Used by \code{downscale.raw}, \code{downscale.delta}, and \code{downscale.deltahybrid}}
\usage{
applyDeltas(obs.hist.daily, obs.hist.monthly, delta_ts, ppt_fun, sigmaN = 6,
  do_checks = FALSE)
}
\description{
Add/multiply deltas to historic daily data to generate future daily SOILWAT2-formatted weather.
Used by \code{downscale.raw}, \code{downscale.delta}, and \code{downscale.deltahybrid}
}
