% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{applyPPTdelta_detailed}
\alias{applyPPTdelta_detailed}
\title{Add/multiply deltas to historic daily precipitation to generate future daily precipitation with checks}
\usage{
applyPPTdelta_detailed(m, data, ydelta, add_days, mult_days, daily, monthly)
}
\arguments{
\item{m}{An integer vector. Each element corresponds to a day (i.e., length(m) is 365
or 366 days) and the values are the number of the month.}

\item{data}{A numeric vector. Precipitation of each day.}

\item{ydelta}{A numeric vector. Delta values for each day. If computed deltas are
monthly, then they must be repeated for each day before passed as argument to this
function.}

\item{add_days}{A logical vector. \code{TRUE} for each day for which \code{ydelta} is
applied additively.}

\item{mult_days}{A logical vector. \code{TRUE} for each day for which \code{ydelta} is
applied multiplicatively.}

\item{daily}{A list. Each element corresponds to one year of simstartyr:endyr is an
object of class \linkS4class{swWeatherData}.}

\item{monthly}{A numeric matrix. Monthly time-series of observed weather calculated
from \code{daily} for the years simstartyr:endyr.}
}
\value{
A list with two elements
 \describe{
   \item{data}{A copy of \code{data} with adjusted values.}
   \item{PPT_to_remove}{The total amount of precipitation that couldn't be removed from \code{data} due to lack of precipitation.}
 }
}
\description{
Add/multiply deltas to historic daily precipitation to generate future daily precipitation with checks
}
