% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{applyPPTdelta_simple}
\alias{applyPPTdelta_simple}
\title{Add/multiply deltas to historic daily precipitation to generate future daily
 precipitation without checks}
\usage{
applyPPTdelta_simple(m, data, ydelta, add_days, mult_days,
  set_negPPT_to0 = TRUE)
}
\arguments{
\item{m}{An integer vector. Each element corresponds to a day (i.e., length(m) is 365
or 366 days) and the values are the number of the month.}

\item{data}{A numeric vector. Precipitation of each day.}

\item{ydelta}{A numeric vector. Delta values for each day. If computed deltas are
monthly, then they must be repeated for each day before passed as argument to this
function.}

\item{add_days}{A logical vector. \code{TRUE} for each day for which \code{ydelta} is
applied additively.}

\item{mult_days}{A logical vector. \code{TRUE} for each day for which \code{ydelta} is
applied multiplicatively.}

\item{set_negPPT_to0}{A logical value. If \code{TRUE} (default) then force days with
resulting negative values of precipitation to 0 -- thereby introducing a positive bias.}
}
\value{
A copy of \code{data} with adjusted values.
}
\description{
Add/multiply deltas to historic daily precipitation to generate future daily
 precipitation without checks
}
