% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soils_Functions.R
\name{assign_aggregation_soillayers}
\alias{assign_aggregation_soillayers}
\title{Determine which soil layers will be aggregated into two, three, or four layer
aggregations for mean daily aggregation output}
\usage{
assign_aggregation_soillayers(layers_depth, daily_lyr_agg)
}
\description{
Determine which soil layers will be aggregated into two, three, or four layer
aggregations for mean daily aggregation output
}
