% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Miscellaneous_Functions.R
\name{biomass}
\alias{biomass}
\alias{Shrub_ANPP}
\alias{Grass_ANPP}
\title{Biomass equations}
\usage{
Shrub_ANPP(MAP_mm)

Grass_ANPP(MAP_mm)
}
\description{
Biomass equations
}
\references{
Milchunas & Lauenroth 1993 (Fig. 2): Y [g/m2/yr] = c1 * MAP [mm/yr] + c2
}
