% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{calc.ScenarioWeather}
\alias{calc.ScenarioWeather}
\title{Extract climate scenario data and downscale to daily weather data}
\usage{
calc.ScenarioWeather(i, ig, il, gcm, site_id, i_tag, clim_source, use_CF,
  use_NEX, climDB_meta, climDB_files, reqGCMs, reqRCPsPerGCM,
  reqDownscalingsPerGCM, climate.ambient, locations, compression_type, getYears,
  assocYears, sim_time, task_seed, opt_DS, project_paths, resume, verbose,
  print.debug)
}
\description{
Extract climate scenario data and downscale to daily weather data
}
