% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorCalculations.R
\name{calc_AddRequestedSoilLayers}
\alias{calc_AddRequestedSoilLayers}
\title{Add soil layers to data.frame (by interpolation) if layers are not already present}
\usage{
calc_AddRequestedSoilLayers(df_soils, df_soils_use, df_soildepths,
  requested_soil_layers, verbose = FALSE, sl_vars_mean = c("Matricd",
  "GravelContent", "Sand", "Clay", "SoilTemp"), sl_vars_sub = c("EvapCoeff",
  "TranspCoeff", "Imperm"), full_interpolation = FALSE)
}
\arguments{
\item{df_soils}{A data frame with soil data. Column names should match with values of
\code{sl_vars_mean} and \code{sl_vars_sub}.}

\item{df_soils_use}{A named logical vector. Names should correspond to column names of
\code{df_soils}. A value of \code{TRUE} indicates that the corresponding variable of
\code{df_soils} is in use/active and will thus be interpolated/exhausted.}

\item{df_soildepths}{A data frame with soil layer depth data. Names of columns with
soil depth information should be constructed as \code{depth_L} where \code{L} is the
layer number. The layer number in turn is used to identify the columns for
\code{df_soils}.}

\item{requested_soil_layers}{A numeric vector of the soil depths for the final set of
soil layers.}

\item{verbose}{A logical value.}

\item{sl_vars_mean}{A vector of strings representing tags that are identifying those
columns of \code{df_soils} by name for which values will be interpolated. See
\code{method} argument of function \code{\link{add_layer_to_soil}}.}

\item{sl_vars_sub}{A vector of strings representing tags that are identifying those
columns of \code{df_soils} by name for which values will be exhausted. See
\code{method} argument of function \code{\link{add_layer_to_soil}}.
@param full_interpolation A logical value determing if layers not in
\code{requested_soil_layers} should be discarded. If so, also interpolate to the
next greatest layer, if possible.}
}
\value{
A list with with five elements: \describe{
   \item{df_soils}{Updated copy of the input.}
   \item{df_soils_use}{Updated copy of the input.}
   \item{df_soildepths}{Updated copy of the input.}
   \item{requested_soil_layers}{Updated copy of the input.}
   \item{has_changed}{A logical value. \code{TRUE} indicates that the inputs have changed.}
 }
}
\description{
Workhorse function for \code{\link{calc_ExtendSoilDatafileToRequestedSoilLayers}}.
}
