% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_PriorCalculations.R
\name{calc_CalculateBareSoilEvaporationCoefficientsFromSoilTexture}
\alias{calc_CalculateBareSoilEvaporationCoefficientsFromSoilTexture}
\title{Calculate potential bare soil evaporation coefficients per soil layer}
\usage{
calc_CalculateBareSoilEvaporationCoefficientsFromSoilTexture(runIDs_adjust,
  sw_input_soils_use, sw_input_soils, sw_input_soillayers, fnames_in,
  resume = TRUE, verbose = FALSE)
}
\description{
Calculate potential bare soil evaporation coefficients per soil layer
}
