% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{calc_DailyScenarioWeather}
\alias{calc_DailyScenarioWeather}
\title{Extract daily climate scenario data}
\usage{
calc_DailyScenarioWeather(
  ids_ToDo,
  clim_source,
  climDB_meta,
  climDB_files,
  reqGCMs,
  reqRCPsPerGCM,
  reqDownscalingsPerGCM,
  locations,
  compression_type,
  getYears,
  sim_scen_ids,
  dir_out_temp,
  dir_failed,
  fdbWeather,
  resume,
  verbose
)
}
\description{
Extract daily climate scenario data
}
\section{Details}{

This function parallelize over \code{reqGCMs} x \code{reqRCPsPerGCM}
combinations, i.e., data for all \code{locations} are extracted for
one value of \code{reqGCMs} x \code{reqRCPsPerGCM} at a time.
This is good if file handling is slow and memory is not limiting.
However, if data cannot be loaded into memory, then this function
cannot work. In this case, this function would need to be re-written to
additionally loop over chunks of \code{locations}.
}

\section{Notes}{
 This function works only if
\itemize{
  \item \code{sim_time[["future_yrs"]]} contains \var{"dall"} and
  \item \code{reqDownscalingsPerGCM} is \var{"idem"}.
}
}

\seealso{
\code{\link{calc_MonthlyScenarioWeather}}
}
