% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{calc_Days_withLoweredPPT}
\alias{calc_Days_withLoweredPPT}
\title{Distribute precipitation among days without too extreme values}
\usage{
calc_Days_withLoweredPPT(data_N, this_newPPTevent_N, sigmaN,
  this_i_extreme, this_pptToDistribute, seed = NA)
}
\arguments{
\item{data_N}{An integer value. The number of days of the precipitation record to
consider.}

\item{this_newPPTevent_N}{An integer value. The number of days which will received
'spill-over' precipitation.}

\item{sigmaN}{An integer value. The multiplier of \code{this_newPPTevent_N} to
determine the range of days to consider.}

\item{this_i_extreme}{An integer value. The index indicating for which day in the
precipitation record, precipitation is removed and redistributed.}

\item{this_pptToDistribute}{An integer value. The amount of precipitation that was
removed from day \code{this_i_extreme} and is now redistributed to other days.}

\item{seed}{A seed set, \code{NULL}, or \code{NA}. \code{NA} will not affect
the state of the \acronym{RNG}; \code{NULL} will re-initialize the \acronym{RNG};
and all other values are passed to \code{\link{set.seed}}.}
}
\description{
Distribute precipitation among days without too extreme values
}
