% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_PriorCalculations.R
\name{calc_ExtendSoilDatafileToRequestedSoilLayers}
\alias{calc_ExtendSoilDatafileToRequestedSoilLayers}
\title{Add soil layers (by interpolation) if not already present}
\usage{
calc_ExtendSoilDatafileToRequestedSoilLayers(requested_soil_layers,
  runIDs_adjust, sw_input_soillayers, sw_input_soils_use, sw_input_soils,
  fnames_in, verbose = FALSE)
}
\description{
Add soil layers (by interpolation) if not already present
}
