% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{calc_MonthlyScenarioWeather}
\alias{calc_MonthlyScenarioWeather}
\title{Extract monthly climate scenario data and downscale to daily weather data}
\usage{
calc_MonthlyScenarioWeather(
  i,
  ig,
  il,
  gcm,
  site_id,
  i_tag,
  clim_source,
  use_CF,
  use_NEX,
  climDB_meta,
  climDB_files,
  reqGCMs,
  reqRCPsPerGCM,
  reqDownscalingsPerGCM,
  climate.ambient,
  locations,
  compression_type,
  getYears,
  assocYears,
  sim_time,
  task_seed,
  opt_DS,
  project_paths,
  dir_failed,
  resume,
  verbose,
  print.debug
)
}
\description{
Extract monthly climate scenario data and downscale to daily weather data
}
\seealso{
\code{\link{calc_DailyScenarioWeather}}
}
