% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorCalculations.R
\name{calc_RequestedSoilLayers}
\alias{calc_RequestedSoilLayers}
\title{Add soil layers (by interpolation) if not already present and store in
soil data input file}
\usage{
calc_RequestedSoilLayers(
  SFSW2_prj_meta,
  SFSW2_prj_inputs,
  runIDs_adjust,
  keep_old_depth = TRUE,
  verbose = FALSE
)
}
\description{
Add soil layers (by interpolation) if not already present and store in
soil data input file
}
\section{Notes}{
 Splitting existing soil layers into two new layers will
  "exhaust" the values of fields \var{\dQuote{EvapCoeff}},
  \var{\dQuote{TranspCoeff}}, and \var{\dQuote{Imperm}}, i.e., sum remains
  constant; and will "interpolate" the values of the fields
  \var{\dQuote{Matricd}}, \var{\dQuote{GravelContent}}, \var{\dQuote{Sand}},
  \var{\dQuote{Clay}}, \var{\dQuote{SoilTemp}}.
}

