% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{calc_SeedlingMortality}
\alias{calc_SeedlingMortality}
\title{Function to calculate mortality under conditions and checks survival limit}
\usage{
calc_SeedlingMortality(kill.conditions, max.duration.before.kill)
}
\description{
Function to calculate mortality under conditions and checks survival limit
}
