% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{calc_timeSlices}
\alias{calc_timeSlices}
\title{Calculate historical and future simulation time slices}
\usage{
calc_timeSlices(sim_time, tbox)
}
\arguments{
\item{sim_time}{A list with elements \code{future_N}, \code{future_yrs},
\code{DScur_startyr}, and \code{DScur_endyr}.}

\item{tbox}{A data.frame or matrix with two rows \code{start} and
\code{end} and two columns \code{first} and \code{second} describing years
including in a specific climate data source.}
}
\value{
A data.frame with rows for each extraction run-slice and
 four columns 'Run', 'Slice', 'Time', and 'Year'.
}
\description{
Calculate historical and future simulation time slices
}
