% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{calc_TimeToGerminate}
\alias{calc_TimeToGerminate}
\title{Function to estimate time to germinate for each day of a given year and
conditions (temperature, top soil \var{SWP})}
\usage{
calc_TimeToGerminate(
  RYyear,
  Germination_WhileFavorable,
  LengthDays_FavorableConditions,
  RYyear_ForEachUsedDay,
  soilTmeanSnow,
  swp.TopMean,
  TmeanJan,
  param,
  seed = NA
)
}
\arguments{
\item{seed}{A seed set, \code{NULL}, or \code{NA}. \code{NA} will not affect
the state of the \var{RNG}; \code{NULL} will re-initialize the \var{RNG};
and all other values are passed to \code{\link{set.seed}}.}
}
\description{
Function to estimate time to germinate for each day of a given year and
conditions (temperature, top soil \var{SWP})
}
