% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{calculate_SeedlingMortality_ByCondition}
\alias{calculate_SeedlingMortality_ByCondition}
\title{Function to calculate mortality under conditions and checks survival limit}
\usage{
calculate_SeedlingMortality_ByCondition(kill.conditions,
  max.duration.before.kill)
}
\description{
Function to calculate mortality under conditions and checks survival limit
}
