% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{calculate_SuitableGrowthThisYear_UnderCondition}
\alias{calculate_SuitableGrowthThisYear_UnderCondition}
\title{Function to calculate favorable conditions for seedling growth for each day of a given year}
\usage{
calculate_SuitableGrowthThisYear_UnderCondition(favorable.conditions,
  consequences.unfavorable)
}
\description{
Function to calculate favorable conditions for seedling growth for each day of a given year
}
