% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{calculate_TimeToGerminate_modifiedHardegree2006NLR}
\alias{calculate_TimeToGerminate_modifiedHardegree2006NLR}
\title{Function to estimate time to germinate for each day of a given year and conditions
(temperature, top soil \var{SWP})}
\usage{
calculate_TimeToGerminate_modifiedHardegree2006NLR(RYyear,
  Germination_DuringFavorableConditions, LengthDays_FavorableConditions,
  RYyear_ForEachUsedDay, soilTmeanSnow, swp.TopMean, TmeanJan, param,
  seed = NA)
}
\arguments{
\item{seed}{A seed set, \code{NULL}, or \code{NA}. \code{NA} will not affect
the state of the \var{RNG}; \code{NULL} will re-initialize the \var{RNG};
and all other values are passed to \code{\link{set.seed}}.}
}
\description{
Function to estimate time to germinate for each day of a given year and conditions
(temperature, top soil \var{SWP})
}
