% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Status_Trackers.R
\name{check_intracker_design}
\alias{check_intracker_design}
\title{Check whether input tracker design is up-to-date}
\usage{
check_intracker_design(ist)
}
\arguments{
\item{ist}{A data.frame representing an input tracker as generated by function
\code{init_intracker}.}
}
\value{
A logical value.
}
\description{
Check whether input tracker design is up-to-date
}
