% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2_DataAccess.R
\name{check_rSW2_version}
\alias{check_rSW2_version}
\title{Check version of a \pkg{rSOILWAT2} input or output object compared to
declared dependency in the package \var{DESCRIPTION}}
\usage{
check_rSW2_version(object, strict = TRUE)
}
\arguments{
\item{object}{An object of \pkg{rSOILWAT2} classes
\code{\linkS4class{swInputData}} or \code{\linkS4class{swOutput}}.}

\item{strict}{A logical value. If \code{FALSE} and check would fail, then
a warning is issued (and \code{TRUE} is returned nevertheless).}
}
\value{
A logical value.
  Returns \code{TRUE} if version of \code{object} meets at least the minimal
  required \pkg{rSOILWAT2} version -- or if \code{strict} is \code{FALSE}.
  Returns \code{FALSE} otherwise.
}
\description{
Check version of a \pkg{rSOILWAT2} input or output object compared to
declared dependency in the package \var{DESCRIPTION}
}
\seealso{
\code{\link[rSOILWAT2]{check_version}}
}
