% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{check_SuitableGrowthThisYear}
\alias{check_SuitableGrowthThisYear}
\title{Function to calculate favorable conditions for seedling growth for each day
of a given year}
\usage{
check_SuitableGrowthThisYear(favorable.conditions, consequences.unfavorable)
}
\description{
Function to calculate favorable conditions for seedling growth for each day
of a given year
}
