% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB_Check.R
\name{check_weatherDB}
\alias{check_weatherDB}
\title{Checks data in a weather database}
\usage{
check_weatherDB(
  dir_prj,
  fdbWeather,
  used_Site_id_by_dbW = NULL,
  repeats = 2L,
  do_preprocess_tempfiles = TRUE,
  n_cores = 20L,
  startyear = 1979,
  endyear = 2010,
  seed = NA
)
}
\arguments{
\item{dir_prj}{A character string. The directory path the \pkg{rSFSW2}
simulation project.}

\item{fdbWeather}{A character string. The file path of weather database.}

\item{used_Site_id_by_dbW}{A numeric vector. A value of \code{NULL}
will process all sites contained in the weather database. Specific
values will process only those sites.}

\item{repeats}{An integer value. The number of times each weather object
is extracted (repeats > 1 enable comparison of the duplicates).}

\item{do_preprocess_tempfiles}{A logical value. Set to TRUE, for instance,
if a previous run was prematurely aborted.}

\item{seed}{A seed set, \code{NULL}, or \code{NA}. \code{NA} will not affect
the state of the \var{RNG}; \code{NULL} will re-initialize the \var{RNG};
and all  other values are passed to \code{\link{set.seed}}.}
}
\description{
Checks data in a weather database
}
