% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{clean_SFSW2_cluster}
\alias{clean_SFSW2_cluster}
\title{Clean memory and workspace of parallel cluster workers}
\usage{
clean_SFSW2_cluster(opt_parallel)
}
\description{
Clean memory and workspace of parallel cluster workers
}
\section{Notes}{
 Do not call 'ls(all = TRUE)' because there are important (hidden)
   \code{.X} objects that are important for proper worker functioning!
}

