% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_parallel.R
\name{clean_SWSF_cluster}
\alias{clean_SWSF_cluster}
\title{Clean the parallel cluster used for a rSWSF simulation project}
\usage{
clean_SWSF_cluster(parallel_backend = c("mpi", "cluster"), cl = NULL,
  verbose = FALSE)
}
\description{
Clean the parallel cluster used for a rSWSF simulation project
}
