% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{clean_worker}
\alias{clean_worker}
\title{Remove almost all objects from a worker's global environment}
\usage{
clean_worker()
}
\description{
Remove all objects except those with a name starting with a dot '.'
}
\section{Notes}{
 Do not call 'ls(all = TRUE)' because there are important (hidden)
   \code{.X} objects that are important for proper worker functioning!
}

