% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{climscen_determine_sources}
\alias{climscen_determine_sources}
\title{Determine climate scenario data sources}
\usage{
climscen_determine_sources(
  climDB_metas,
  xmain,
  runIDs_sites,
  runsN_sites,
  how_determine_sources = c("SWRunInformation", "order"),
  scen_sources,
  fmain,
  fpreprocin
)
}
\description{
Allow for multiple data sources among sites but not multiple sources per site
(for that you need a new row in the \var{\sQuote{InputMain}} spreadsheet)
}
