% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testproject_Functions.R
\name{compare_test_output}
\alias{compare_test_output}
\title{Compare test project output database with reference database}
\usage{
compare_test_output(dir_test, dir_ref = NULL, verbose = FALSE)
}
\arguments{
\item{dir_test}{A character string. Path to test project folder.}

\item{dir_ref}{A character string. Path to folder with reference database.}

\item{verbose}{A logical value. If \code{TRUE} then messages are printed.}
}
\value{
The output from \code{\link{compare_two_dbOutput}}.
}
\description{
Reference database is identified by containing \code{basename(dir_test)} in
the file name.
}
\seealso{
\code{\link{compare_two_dbOutput}}
}
