% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{compare_two_dbOutput}
\alias{compare_two_dbOutput}
\title{Compare one output database with another output database}
\usage{
compare_two_dbOutput(dbOut1, dbOut2, tol = 0.001, comp_absolute = TRUE,
  verbose = FALSE)
}
\arguments{
\item{dbOut1}{A character string. Path to first output database used as
reference.}

\item{dbOut2}{A character string. Path to second output database.}

\item{tol}{A numeric value. Differences smaller than tolerance are not
reported. Passed to \code{\link{all.equal}}.}

\item{comp_absolute}{A logical value. If \code{TRUE} then absolute
comparisons will be reported, otherwise relative differences. See argument
\code{scale} of function \code{\link{all.equal}}.}

\item{verbose}{A logical value. If \code{TRUE} then messages are printed.}
}
\value{
A (possibly empty) list of character vectors describing differences
  between \code{dbOut2} and \code{dbOut1} databases.
}
\description{
Compare one output database with another output database
}
\seealso{
\code{\link{all.equal}}
}
