% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{controlExtremePPTevents}
\alias{controlExtremePPTevents}
\title{Check for and handle days with too extreme precipitation values}
\usage{
controlExtremePPTevents(data, dailyPPTceiling, sigmaN, do_checks = FALSE,
  mfact = 10)
}
\arguments{
\item{data}{A numeric vector. Daily values of precipitation.}

\item{dailyPPTceiling}{A numeric value. The maximum value of daily precipitation.
Values above this limit will be removed and redistributed to other days.}

\item{sigmaN}{An integer value. A multiplicator of \code{stats::sd} for data checks.}

\item{do_checks}{A logical value. See details.}

\item{mfact}{A numeric value. See details.}
}
\description{
Check for and handle days with too extreme precipitation values
}
\details{
If \code{do_check == TRUE} and any daily precipitation is equal or larger than
 \code{mfact * dailyPPTceiling}, then the code will error out.
}
