% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{convert_temperature}
\alias{convert_temperature}
\title{Converts units of temperature data}
\usage{
convert_temperature(x, unit_from, unit_to = "C")
}
\arguments{
\item{x}{A numeric vector. Temperature data as monthly series in units of
\code{unit_from}.}

\item{unit_from}{A character string. Units of data in \code{x}. Currently,
supported units include "K", "F", and "C".}

\item{unit_to}{A character string. Units to which data are converted.
Currently, supported unit is "C".}
}
\value{
A numeric vector of the same size as \code{x} in units of
  \code{unit_to}.
}
\description{
Converts units of temperature data
}
