% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{create_dbWork}
\alias{create_dbWork}
\title{Create a \var{SQLite}-database \code{dbWork} to manage runs of a \pkg{rSFSW2}
simulation project}
\usage{
create_dbWork(path, jobs)
}
\arguments{
\item{path}{A character string. Path to the folder where the database will be
created.}

\item{jobs}{An integer matrix. Each row corresponds to one call of the
simulation function \code{do_OneSite}, i.e., \code{runsN_main} x
\code{expN}. The columns \code{runID_total}, \code{runID_sites},
\code{include_YN} represent a running ID, the \code{site_id} (row number in
main input file), and a flag whether site is being simulated or not. See
\code{\link{indices}}.}
}
\value{
Invisibly \code{TRUE}
}
\description{
Create a \var{SQLite}-database \code{dbWork} to manage runs of a \pkg{rSFSW2}
simulation project
}
