% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_databases.R
\name{dbConnect2}
\alias{dbConnect2}
\title{Connect to SQLite-database}
\usage{
dbConnect2(dbname, flags = RSQLite::SQLITE_RW, verbose = FALSE,
  repeats = 10L, sleep_s = 5, seed = NA)
}
\arguments{
\item{dbname}{A character string. The path including name to the database.}

\item{flags}{An integer value. See \code{\link[RSQLite]{dbConnect}}. Defaults to
read/write mode.}

\item{verbose}{A logical value.}

\item{repeats}{An integer value. The maximum number of failed attempts to execute the
\code{SQL} statement(s) if the database is locked.}

\item{sleep_s}{A numeric value. The average number to sleep before attempting to
execute the \code{SQL} statement(s) if the database was locked.}

\item{seed}{An R object. Passed to \code{\link{set.seed}} if not \code{NA}.}
}
\value{
A \linkS4class{SQLiteConnection} object on success; an object of class
  \code{\link[base:try]{try-error}} on failure.
}
\description{
A wrapper around \code{\link[RSQLite]{dbConnect}} that catches errors and attempts
to connect up to \code{repeats} before giving up.
}
\seealso{
\code{\link[RSQLite]{dbConnect}}
}
