% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{dbOut_read_variables_from_scenario}
\alias{dbOut_read_variables_from_scenario}
\title{Reads variables/fields from the header view or from tables including sites,
runs and one of the overall aggregation tables for one of the scenarios
and optionally subsets rows by a where-clause}
\usage{
dbOut_read_variables_from_scenario(fname_dbOut, variables = NULL,
  MeanOrSD = c("Mean", "SD"), scenario = "Current",
  whereClause = NULL)
}
\arguments{
\item{fname_dbOut}{A character string. The path to the output database.}

\item{variables}{A vector of character strings. The (partial) names of
variables/columns/fields to be read.}

\item{MeanOrSD}{A character string. Identify which type of overall aggregated
table to read.}

\item{scenario}{A character string. This must be one of the values from the
\var{Scenario} values from the \var{header} view.}

\item{whereClause}{A character string. This must be of the structure
\var{field_name='value'} or \code{NULL}}
}
\value{
A \code{data.frame}
}
\description{
Reads variables/fields from the header view or from tables including sites,
runs and one of the overall aggregation tables for one of the scenarios
and optionally subsets rows by a where-clause
}
