% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{dbOutput_print_scenarioList}
\alias{dbOutput_print_scenarioList}
\title{Query \code{dbOutput} for scenario table and write spreadsheet to disk}
\usage{
dbOutput_print_scenarioList(dbOutput, fnameScenarioList)
}
\description{
Query \code{dbOutput} for scenario table and write spreadsheet to disk
}
\examples{
dbOutput_print_scenarioList(
  SFSW2_prj_meta[["fnames_out"]][["dbOutput"]],
  fnameScenarioList = "Table_ScenarioDescription.csv"
)

}
\seealso{
\code{\link[=write_scen_description]{write_scen_description()}}
}
