% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{dbOutput_subset}
\alias{dbOutput_subset}
\title{Make a copy of \var{\code{dbOutput}} with a subset of tables and/or fields}
\usage{
dbOutput_subset(
  dbOut_fname,
  dbNew_fname,
  fields_include = NULL,
  fields_exclude = NULL,
  subset_scenarios = NULL,
  subset_experiments = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dbOut_fname}{A character string. The file path of the main
\var{\code{dbOutput}}.}

\item{dbNew_fname}{A character string. The file path and name of the new
copy of the database with a subset of the values of \var{\code{dbOutput}}.}

\item{fields_include}{A named list of vectors with character strings. The
field names per table to be selected. Each table is represented by a
correspondingly named element. If \code{NULL}, then all tables and all
fields are included. If a named element is \code{NULL}, then all fields of
the corresponding table are included. Except for those fields listed in
\code{fields_exclude}.}

\item{fields_exclude}{A named list of vectors with character strings. The
field names per table to be excluded from the subset. Each table is
represented by a correspondingly named element. If \code{NULL}, then no
fields are excluded from the subset operation.}

\item{subset_scenarios}{A vector of character strings. If not \code{NULL},
then only records are copied with a \var{P_id} that belongs to an
included scenario.}

\item{subset_experiments}{A vector of character strings. If not \code{NULL},
then only records are copied with a \var{P_id} that belongs to an
included experimental treatment label. This and the \code{subset_scenarios}
condition are combined with a logical \var{AND}.}

\item{verbose}{A logical value.}
}
\description{
The copy includes all design tables and their full content and a subset
of the records and/or fields of the output tables.
}
