% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB_Check.R
\name{dbW_has_missingClimScens}
\alias{dbW_has_missingClimScens}
\title{Determine which site_ids in the weather database do not have weather data for every
requested climate scenario}
\usage{
dbW_has_missingClimScens(fdbWeather, SFSW2_prj_inputs, runIDs_sites, req_scenN,
  opt_verbosity, opt_chunks)
}
\value{
A logical vector of length \code{SFSW2_prj_meta[["sim_size"]][["runsN_master"]]}.
 The elements at location \code{SFSW2_prj_meta[["sim_size"]][["runIDs_sites"]]} are
 \code{TRUE} indicating the site_id for which the weather database lacks all data. If
 all data are in the weather database, then \code{FALSE}.
}
\description{
Determine which site_ids in the weather database do not have weather data for every
requested climate scenario
}
