% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB_Check.R
\name{dbW_missingAmbient_siteIDs}
\alias{dbW_missingAmbient_siteIDs}
\title{Determine which site_ids in the weather database do not have weather data for the
ambient climate scenario}
\usage{
dbW_missingAmbient_siteIDs(req_sites, fdbWeather, dbW_iSiteTable = NULL,
  opt_chunks, verbose = FALSE)
}
\description{
Determine which site_ids in the weather database do not have weather data for the
ambient climate scenario
}
\section{Notes}{
 The ambient climate scenario is supposed to always be scenario 1;
 see how \code{setup_scenarios} constructs the scenarios.
}

