% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB_Check.R
\name{dbW_sites_with_missingClimScens}
\alias{dbW_sites_with_missingClimScens}
\title{Determine which site_ids in the weather database do not have weather data for every
requested climate scenario}
\usage{
dbW_sites_with_missingClimScens(fdbWeather, site_labels = NULL,
  siteID_by_dbW = NULL, scen_labels = NULL, scenID_by_dbW = NULL,
  chunk_size = 500L, verbose = FALSE)
}
\value{
A logical vector of length \code{site_labels} respectively \code{siteID_by_dbW}.
 The \code{TRUE} elements indicate sites for which the weather database lacks some or
 all data. If all data are in the weather database, then \code{FALSE}.
}
\description{
Determine which site_ids in the weather database do not have weather data for every
requested climate scenario
}
