% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_check}
\alias{dbWork_check}
\title{Check run status}
\usage{
dbWork_check(path, runIDs)
}
\arguments{
\item{path}{A character string. Path to the folder where the database will be created.}

\item{runIDs}{An integer vector. The identification numbers of queried runs,
i.e., values out of \code{runIDs_total}, see \code{\link{indices}}.}
}
\value{
A data.frame with three columns 'completed', 'failed', and 'inwork'
}
\description{
Check run status
}
