% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_check_granular}
\alias{dbWork_check_granular}
\title{Check granular run status}
\usage{
dbWork_check_granular(path, runIDs)
}
\arguments{
\item{path}{A character string. Path to the folder where the database will be
created.}

\item{runIDs}{An integer vector. The identification numbers of queried runs,
i.e., values out of \code{runIDs_total}, see \code{\link{indices}}.}
}
\value{
A data.frame with columns as created by
  \code{\link{add_granular_dbWork}} and one row per \code{runIDs}.
}
\description{
Check granular run status
}
