% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_clean}
\alias{dbWork_clean}
\title{Do maintenance work on a SQLite-database \code{dbWork} of a rSFSW2 simulation project}
\usage{
dbWork_clean(path)
}
\arguments{
\item{path}{A character string. Path to the folder where the database will be created.}
}
\value{
A logical value
}
\description{
Some code, power, and system failures may leave \code{dbWork} in an incomplete state,
e.g., a rollback journal is present, or runs are marked as 'inwork' even though no
runs are currently being worked on. This function cleans such situations up.
}
