% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_Ntodo}
\alias{dbWork_Ntodo}
\title{Numbers of runs of a \pkg{rSFSW2} simulation project which are uncompleted and
not \var{\dQuote{inwork}}}
\usage{
dbWork_Ntodo(path)
}
\arguments{
\item{path}{A character string. Path to the folder where the database will be created.}
}
\value{
An integer value.
}
\description{
Numbers of runs of a \pkg{rSFSW2} simulation project which are uncompleted and
not \var{\dQuote{inwork}}
}
