% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_redo}
\alias{dbWork_redo}
\title{Set runs information that need to be redone / simulated (again)}
\usage{
dbWork_redo(path, runIDs)
}
\arguments{
\item{path}{A character string. Path to the folder where the database will be created.}
}
\value{
A logical vector indicating success.
}
\description{
Set runs information that need to be redone / simulated (again)
}
