% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_timing}
\alias{dbWork_timing}
\title{Extract stored execution times of completed runs of a rSFSW2 simulation project}
\usage{
dbWork_timing(path)
}
\value{
A numeric vector of execution time in seconds.
}
\description{
Extract stored execution times of completed runs of a rSFSW2 simulation project
}
