% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_dbWork.R
\name{dbWork_todos}
\alias{dbWork_todos}
\title{Extract identification numbers of runs of a SWSF simulation project which are
 uncompleted and not \code{inwork}}
\usage{
dbWork_todos(path)
}
\value{
An integer vector of \code{runIDs}.
}
\description{
Extract identification numbers of runs of a SWSF simulation project which are
 uncompleted and not \code{inwork}
}
