% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_update_granular}
\alias{dbWork_update_granular}
\title{Update granular run information of a \pkg{rSFSW2} simulation project}
\usage{
dbWork_update_granular(path, table, Pid, status, verbose = FALSE)
}
\arguments{
\item{path}{A character string. Path to the folder where the database will be
created.}

\item{table}{A character string.}

\item{Pid}{An integer value. The identification number of the current output,
i.e., see \code{\link{indices}}.}

\item{status}{A logical value. \code{FALSE} indicates "no longer needed",
i.e., that the output for \code{table} has successfully been generated (see
\code{\link{add_granular_dbWork}}).}

\item{verbose}{A logical value.}
}
\value{
A logical value whether the status was successfully updated.
}
\description{
Update granular run information of a \pkg{rSFSW2} simulation project
}
