% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_dbWork.R
\name{dbWork_update_job}
\alias{dbWork_update_job}
\title{Update run information of a SWSF simulation project}
\usage{
dbWork_update_job(path, runID, status = c("completed", "failed", "inwork"),
  time_s = "NULL", with_filelock = NULL, verbose = FALSE)
}
\arguments{
\item{runID}{An integer value. The identification number of the current run,
i.e., a value out of \code{runIDs_total}, see \code{\link{indices}}.}

\item{status}{A character string. One of "completed", "failed", "inwork".}

\item{time_s}{A numeric value. The execution time in seconds; used if \code{status} is one of
"completed" and "failed".}

\item{with_filelock}{A character string. The file path for locking access to
\code{dbWork} with a file lock, i.e., to provide
synchronization during parallel processing. If \code{NULL}, no file locking is used.}

\item{verbose}{A logical value. If \code{TRUE}, status messages about file lock and
database access are printed}
}
\value{
A logical value whether the status was successfully updated.
}
\description{
Update run information of a SWSF simulation project
}
