% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_update_status}
\alias{dbWork_update_status}
\title{Update modification status of a \pkg{rSFSW2} simulation project}
\usage{
dbWork_update_status(path, status, verbose = FALSE)
}
\arguments{
\item{path}{A character string. Path to the folder where the database will be
created.}

\item{status}{A logical value. \code{FALSE} indicates "not modified", i.e.,
that \code{dbWork} and generated output \code{dbOutput} and
\code{dbTempOut} is synchronized; \code{TRUE} indicates "modified".}

\item{verbose}{A logical value.}
}
\value{
A logical value whether the status was successfully updated.
}
\description{
Update modification status of a \pkg{rSFSW2} simulation project
}
\seealso{
\code{\link{add_status_dbWork}}
}
