% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testproject_Functions.R
\name{delete_test_output}
\alias{delete_test_output}
\title{Delete output of a test project}
\usage{
delete_test_output(dir_test, delete_filepaths = NULL)
}
\arguments{
\item{dir_test}{A character string. Path to overall test project folder.}

\item{delete_filepaths}{A vector of character strings or \code{NULL}. Files
to delete.}
}
\description{
Delete output of a test project
}
