% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indices.R
\name{determine_simulation_size}
\alias{determine_simulation_size}
\title{Calculate the size of the simulation experiment: number of runs, etc.}
\usage{
determine_simulation_size(SWRunInformation, include_YN, sw_input_experimentals,
  sim_scens)
}
\description{
Calculate the size of the simulation experiment: number of runs, etc.
}
\section{Details of indices}{
 \code{\link{indices}}.
}

\seealso{
\code{\link{indices}}, \code{\link{it_exp}}, \code{\link{it_site}},
 \code{\link{it_Pid}}
}
