% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{dir_safe_create}
\alias{dir_safe_create}
\title{Create the elements of paths}
\usage{
dir_safe_create(paths, showWarnings = FALSE, recursive = TRUE,
  mode = "0777")
}
\arguments{
\item{paths}{A list or vector of strings. Path names to be created.}

\item{showWarnings}{logical; should the warnings on failure be shown?}

\item{recursive}{logical. Should elements of the path other than the
    last be created?  If true, like the Unix command \command{mkdir -p}.}

\item{mode}{the mode to be used on Unix-alikes: it will be
    coerced by \code{\link{as.octmode}}.  For \code{Sys.chmod} it is
    recycled along \code{paths}.}
}
\value{
An invisible list of length equal to the length of \code{paths}. The elements
  are \code{NULL} for invalid elements of \code{paths}, a logical value for the
  elements of \code{paths} with a successful calls to \code{\link{dir.create}}, or
  an object of class \code{try-error} for a failed call to \code{\link{dir.create}}.
}
\description{
This is a wrapper for the function \code{\link{dir.create}} using different default
settings and allowing multiple path names as input. The function checks if the
provided paths may be valid names and catches any other errors with \code{try}.
}
\seealso{
\code{\link{dir.create}}
}
