% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_MeanMonthlyClimate.R
\name{do_ExtractSkyDataFromNCEPCFSR_Global}
\alias{do_ExtractSkyDataFromNCEPCFSR_Global}
\title{Extract gridded mean monthly data from NCEP/CFSR for sites globally}
\usage{
do_ExtractSkyDataFromNCEPCFSR_Global(MMC, SWRunInformation, SFSW2_prj_meta,
  opt_parallel, opt_chunks, resume, verbose)
}
\description{
Extract gridded mean monthly data from NCEP/CFSR for sites globally
}
\section{Monthly data}{
 (http://rda.ucar.edu/datasets/ds093.2/):
 ds093.2 - NCEP Climate Forecast System Reanalysis (CFSR) Monthly Products, January
 1979 to December 2010, 0.313-deg: monthly mean (4 per day) of forecasts of 6-hour
 average.
 \describe{
   \item{relative humidity}{percentage for entire atmosphere at 2 m above ground
       [0.5-deg]: 'pgbh06.gdas.R_H.2m.grb2' --> means for Jan-Dec}
   \item{wind (m s-1)}{u- and v-component at 10 m above ground in m s-1:
       'flxf06.gdas.WND.10m.grb2' (u- and v-component) --> means for Jan-Dec}
   \item{total cloud cover}{percentage of entire atmosphere as a single layer:
       'flxf06.gdas.T_CDC.EATM.grb2' --> means for Jan-Dec}
 }
}

\references{
Environmental Modeling Center/National Centers for Environmental
 Prediction/National Weather Service/NOAA/U.S. Department of Commerce. 2010. NCEP
 Climate Forecast System Reanalysis (CFSR) Monthly Products, January 1979 to December
 2010. Research Data Archive at the National Center for Atmospheric Research,
 Computational and Information Systems Laboratory.
 http://rda.ucar.edu/datasets/ds093.2/. Accessed 8 March 2012.
}
