% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{do_ExtractSoilDataFromISRICWISE_Global}
\alias{do_ExtractSoilDataFromISRICWISE_Global}
\title{Extract soil data from one of the ISRIC-WISE datasets}
\usage{
do_ExtractSoilDataFromISRICWISE_Global(MMC, sim_size, sim_space, dir_ex_soil,
  fnames_in, dataset = c("ISRICWISEv12", "ISRICWISE30secV1a"), opt_parallel,
  resume, verbose)
}
\arguments{
\item{dataset}{A character string. Identifies the ISRIC-WISE dataset from which
soil data are extracted. See details.}
}
\description{
Extract soil data from one of the ISRIC-WISE datasets
}
\section{Details}{
 \code{dataset} is current implemented for values: \describe{
   \item{"ISRICWISEv12"}{Dataset 'ISRIC-WISE 5-arcmin v1.2' (Batjes 2012)}
   \item{"ISRICWISE30secV1a"}{Dataset 'ISRIC-WISE 30-arcsec v1.0' (Batjes 2015, 2016)}
 }
}

\section{Note}{
 Cells with no soil values include those with \code{SUID = c(0 = Water,
 6997 = Water, 6694 = Rock, or 6998 = Glacier)}
}

\references{
Batjes, N. H. 2012. ISRIC-WISE derived soil properties on a 5 by 5
 arc-minutes global grid (ver. 1.2). Report 2012/01 (with data set, available at
 www.isric.org). ISRIC-World Soil Information, Wageningen, The Netherlands.
 http://www.isric.org/data/isric-wise-derived-soil-properties-5-5-arc-minutes-global-grid-version-12

Batjes N.H. 2016. Harmonised soil property values for broad-scale
  modelling (WISE30sec) with estimates of global soil carbon stocks. Geoderma 269,
  61-68 (http://dx.doi.org/10.1016/j.geoderma.2016.01.034).

Batjes N.H. 2015. World soil property estimates for broad-scale modelling
  (WISE30sec, ver. 1.0). Report 2015/01, ISRIC—World Soil Information, Wageningen
  [available at ISRIC Soil Data Hub](http://geonode.isric.org/search/?title__icontains=World%20soil%20property%20estimates%20for%20broad-scale%20modelling%20(WISE30sec)&limit=100&offset=0),
  with addendum and corrigendum.
}
