% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{do_ExtractSoilDataFromISRICWISEv12_Global}
\alias{do_ExtractSoilDataFromISRICWISEv12_Global}
\title{Extract soil data from ISRIC-WISE 5-arcmin v1.2}
\usage{
do_ExtractSoilDataFromISRICWISEv12_Global(MMC, sim_size, sim_space, dir_ex_soil,
  fnames_in, opt_parallel, resume, verbose)
}
\description{
Extract soil data from ISRIC-WISE 5-arcmin v1.2
}
\section{Note}{
 Cells with no soil values include those with \code{SUID = c(0 = Water,
 6997 = Water, 6694 = Rock, or 6998 = Glacier)}
}

\references{
Batjes, N. H. 2012. ISRIC-WISE derived soil properties on a 5 by 5
 arc-minutes global grid (ver. 1.2). Report 2012/01 (with data set, available at
 www.isric.org). ISRIC-World Soil Information, Wageningen, The Netherlands.
 http://www.isric.org/data/isric-wise-derived-soil-properties-5-5-arc-minutes-global-grid-version-12
}
