% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{do_ExtractSoilDataFromSSURGO}
\alias{do_ExtractSoilDataFromSSURGO}
\title{Download and extract SSURGO data}
\usage{
do_ExtractSoilDataFromSSURGO(MMC, dir_data_SSURGO, fnames_in, verbose,
  print.debug = FALSE, resume, SWRunInformation, sim_size, sim_space,
  dir_ex_soil, do_interpolate, req_soil_lyrs)
}
\description{
Download the respective SSURGO area for each site, extract
the data into both spatial and tabular files, choose the most prominent keys,
extract the respective data within the keys, and populate the soil layers and soil texture
slots in the MMC variable with that data.

Prints out why a function call failed
}
\note{
\preformatted{Gridded STATSGO data is extracted for sites with the following issues:
1. The site failed to download or extract
2. Keys could not be chosen due to incomplete data
3. No horizons existed
4. The soil texture lacked data for any given layer of a site}

Will also enable the extraction of STATSGO once all sites are finished
}
\author{
Zachary Kramer, \email{kramer.zachary.nau@gmail.com}
}
