% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Run.R
\name{do_OneSite}
\alias{do_OneSite}
\title{The main simulation function which does all the heavy lifting}
\usage{
do_OneSite(
  i_sim,
  i_SWRunInformation,
  i_sw_input_soillayers,
  i_sw_input_treatments,
  i_sw_input_cloud,
  i_sw_input_prod,
  i_sw_input_site,
  i_sw_input_soils,
  i_sw_input_swrcp,
  i_sw_input_weather,
  i_sw_input_climscen,
  i_sw_input_climscen_values,
  SimParams,
  fdbWeather_by_scen
)
}
\description{
The main simulation function which does all the heavy lifting
}
\details{
For contributors only: This function cannot return prematurely because
 (i.e., don't use \code{return}); otherwise the management of simulation runs will
 fail. If a condition is met that prevents proper continuation/execution of a
 simulation, then the appropriate element in the variable \code{tasks} should be set
 to 0. The variable \code{tasks} contains the elements \code{create}, \code{execute},
 and \code{aggregate} with the values: \itemize{
   \item -1 indicates "don't do" a task element
   \item 0 indicates that the task element has/is "failed"
   \item 1 indicates "to do" a task element
   \item 2 indicates that a task element had "success" in executing relevant code
 }
}
