% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{downscale.periods}
\alias{downscale.periods}
\title{Time periods for downscaling functions}
\usage{
downscale.periods(obs.hist.daily, obs.hist.monthly,
  scen.hist.monthly = NULL, scen.fut.monthly = NULL, years = NULL,
  DScur_startyear = NULL, DScur_endyear = NULL,
  DSfut_startyear = NULL, DSfut_endyear = NULL)
}
\arguments{
\item{obs.hist.daily}{A list. Each element corresponds to one year of
\code{simstartyr:endyr} is an object of
\code{\link[rSOILWAT2:swWeatherData-class]{rSOILWAT2::swWeatherData}}.}

\item{obs.hist.monthly}{A numeric matrix. Monthly time-series of observed weather
calculated from \code{obs.hist.daily} for the years \code{simstartyr:endyr}.}

\item{scen.hist.monthly}{A numeric matrix. Monthly time-series of scenario weather
during the historic time period \code{DScur_startyr:DScur_endyr}}

\item{scen.fut.monthly}{A numeric matrix. Monthly time-series of scenario weather
during the projected time period \code{DSfut_startyr:DSfut_endyr}}
}
\description{
Time periods for downscaling functions
}
