% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_ClimateDownscaling.R
\name{downscale.raw}
\alias{downscale.raw}
\title{Downscale with the 'direct approach'}
\usage{
downscale.raw(obs.hist.daily, obs.hist.monthly, scen.fut.monthly, itime,
  years = NULL, sim_time = NULL, opt_DS = list(ppt_type = "detailed",
  sigmaN = 6, PPTratioCutoff = 10), dailyPPTceiling, do_checks = TRUE, ...)
}
\arguments{
\item{obs.hist.daily}{A list. Each element corresponds to one year of simstartyr:endyr is an object of class \linkS4class{swWeatherData}.}

\item{obs.hist.monthly}{A numeric matrix. Monthly time-series of observed weather calculated from \code{obs.hist.daily} for the years simstartyr:endyr.}

\item{scen.fut.monthly}{A numeric matrix. Monthly time-series of scenario weather during the projected time period DSfut_startyr:DSfut_endyr}

\item{opt_DS}{A named list.}

\item{do_checks}{A logical value. If \code{TRUE} perform several sanity checks on the data.}
}
\description{
See 'direct' approach in Lenderink et al. (2007)
}
\references{
Lenderink, G., A. Buishand, and W. van Deursen. 2007. Estimates of future discharges of the river Rhine using two scenario methodologies: direct versus delta approach. Hydrology and Earth System Sciences 11:1145-1159.
}
