% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{downscale}
\alias{downscale}
\title{Downscale and temporal disaggregation}
\arguments{
\item{obs.hist.daily}{A list. Each element corresponds to one year of simstartyr:endyr
is an object of class \linkS4class{swWeatherData}.}

\item{daily}{A list. Each element corresponds to one year of simstartyr:endyr is an
object of class \linkS4class{swWeatherData}.}

\item{obs.hist.monthly}{A numeric matrix. Monthly time-series of observed weather
calculated from \code{obs.hist.daily} for the years simstartyr:endyr.}

\item{monthly}{A numeric matrix. Monthly time-series of observed weather calculated
from \code{daily} for the years simstartyr:endyr.}

\item{scen.hist.monthly}{A numeric matrix. Monthly time-series of scenario weather
during the historic time period DScur_startyr:DScur_endyr}

\item{scen.fut.monthly}{A numeric matrix. Monthly time-series of scenario weather
during the projected time period DSfut_startyr:DSfut_endyr}

\item{opt_DS}{A named list.}

\item{do_checks}{A logical value. If \code{TRUE} perform several sanity checks on the
data.}
}
\description{
Downscale and temporal disaggregation
}
\section{Details}{
 Units are [degree Celsius] for temperature and [cm / day] or
 [cm / month], respectively, for precipitation
}

