% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mathematical_Functions.R
\name{erf}
\alias{erf}
\title{Error function}
\usage{
erf(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of the size of \code{x}.
}
\description{
Error function
}
\seealso{
Code is from examples of \code{\link[stats]{pnorm}}.
}
