% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vegetation.R
\name{estimate_PotNatVeg_composition}
\alias{estimate_PotNatVeg_composition}
\title{Calculate the composition of a potential natural vegetation based of shrub,
C3 grass, and C4 grass functional group components using climate
relationships}
\usage{
estimate_PotNatVeg_composition(MAP_mm, MAT_C, mean_monthly_ppt_mm,
  dailyC4vars, isNorth = TRUE, shrub_limit = 0.2, fix_annuals = TRUE,
  Annuals_Fraction = 0, fix_C4grasses = FALSE, C4_Fraction = NA,
  fix_C3grasses = FALSE, C3_Fraction = NA, fix_shrubs = FALSE,
  Shrubs_Fraction = NA, fix_forbs = FALSE, Forbs_Fraction = NA,
  fix_BareGround = TRUE, BareGround_Fraction = 0)
}
\arguments{
\item{MAP_mm}{A numeric value. Mean annual precipitation in millimeter of the
location.}

\item{MAT_C}{A numeric value. Mean annual temperature in degree Celsius.}

\item{mean_monthly_ppt_mm}{A numeric vector of length 12. Mean monthly
precipitation in millimeter.}

\item{dailyC4vars}{A named numeric vector of length 3. \describe{
\item{\code{Month7th_NSadj_MinTemp_C}}{Mean minimum temperature of July on
  the northern hemisphere and January on the southern hemisphere}
\item{\code{DegreeDaysAbove65F_NSadj_DaysC}}{Degree days above 65 F = 18.33
  C in units of days x degree Celsius}
\item{\code{LengthFreezeFreeGrowingPeriod_NSadj_Days}}{Mean annual number
  of days of the longest continuous period where minimum daily temperature
  remain above freezing} }}

\item{isNorth}{A logical value. \code{TRUE} for locations on northern
hemisphere.}

\item{shrub_limit}{A numeric value. Default value is 0.2 based on page 1213
of Paruelo & Lauenroth 1996.}

\item{fix_annuals}{A logical value. If \code{TRUE}, then value for the annual
component is fixed at \code{Annuals_Fraction}.}

\item{Annuals_Fraction}{A numeric value. Default value is 0. A value between
0 and 1.}

\item{fix_C4grasses}{A logical value. If \code{TRUE}, then value for the
C4-grass component is fixed at \code{C4_Fraction} instead of calculated
from climatic relationships.}

\item{C4_Fraction}{A numeric value between 0 and 1. \code{NA} is treated as
if \code{fix_C4grasses} is \code{FALSE}.}

\item{fix_C3grasses}{A logical value. If \code{TRUE}, then value for the
C3-grass component is fixed at \code{C3_Fraction} instead of calculated
from climatic relationships.}

\item{C3_Fraction}{A numeric value between 0 and 1. \code{NA} is treated as
if \code{fix_C3grasses} is \code{FALSE}.}

\item{fix_shrubs}{A logical value. If \code{TRUE}, then value for the shrub
component is fixed at \code{Shrubs_Fraction} instead of calculated from
climatic relationships.}

\item{Shrubs_Fraction}{A numeric value between 0 and 1. \code{NA} is treated
as if \code{fix_shrubs} is \code{FALSE}.}

\item{fix_forbs}{A logical value. If \code{TRUE}, then value for the forb
component is fixed at \code{Forbs_Fraction}.}

\item{Forbs_Fraction}{A numeric value. Default value is 0. A value between 0
and 1.}

\item{fix_BareGround}{A logical value. If \code{TRUE}, then value for the
bare ground component is fixed at \code{BareGround_Fraction}.}

\item{BareGround_Fraction}{A numeric value. Default value is 0. A value
between 0 and 1.}
}
\value{
A list with two named numeric vectors. \describe{
  \item{Composition}{Relative composition [0-1] of the vegetation for
    \code{Grasses}, \code{Shrubs}, \code{Trees}, \code{Forbs}, and
    \code{BareGround}. \code{Grasses} are the sum of C3-grasses, C4-grasses,
    and annuals functional groups. \code{Trees} is set to zero. The sum of
    \code{Composition} is 1.}
  \item{grasses.c3c4ann.fractions}{Relative contribution [0-1] of the
    C3-grasses, C4-grasses, and annuals functional groups. The sum of
    \code{grasses.c3c4ann.fractions} is 1.} }
}
\description{
Equations by Paruelo & Lauenroth 1996 are used to estimate shrub, C3 grass,
and C3 grass components. Equations by Teeri & Stowe 1976 are used to limit
occurrence of C4 grasses.
}
\section{Note}{
 This function does not estimate cover of the tree, forb,
  and bare-ground components; instead, these are set to 0.
}

\references{
Paruelo J.M., Lauenroth W.K. (1996) Relative abundance of plant
  functional types in grasslands and shrublands of North America. Ecological
  Applications, 6, 1212-1224.

Teeri J.A., Stowe L.G. (1976) Climatic patterns and the
  distribution of C4 grasses in North America. Oecologia, 23, 1-12.
}
