% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{exit_SFSW2_cluster}
\alias{exit_SFSW2_cluster}
\title{Clean up and terminate a parallel cluster used for a \pkg{rSFSW2} simulation project}
\usage{
exit_SFSW2_cluster(verbose = FALSE)
}
\arguments{
\item{verbose}{A logical value.}
}
\description{
Clean up and terminate a parallel cluster used for a \pkg{rSFSW2} simulation project
}
