% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{export_objects_to_workers}
\alias{export_objects_to_workers}
\title{Export objects to workers}
\usage{
export_objects_to_workers(obj_env, parallel_backend = c("mpi", "socket"),
  cl = NULL)
}
\arguments{
\item{obj_env}{An environment containing R objects to export.}

\item{parallel_backend}{A character vector, either \var{\dQuote{mpi}}' or
\var{\dQuote{socket}}}

\item{cl}{A parallel (socket) cluster object}
}
\value{
A logical value. \code{TRUE} if every object was exported
  successfully.
}
\description{
Export objects to workers
}
