% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB.R
\name{extract_daily_weather_from_livneh}
\alias{extract_daily_weather_from_livneh}
\title{Extract Gridded Weather Data from a Livneh Database}
\usage{
extract_daily_weather_from_livneh(
  dir_data,
  dir_temp,
  site_ids,
  site_ids_by_dbW,
  coords,
  start_year,
  end_year,
  f_check = TRUE,
  backup = TRUE,
  comp_type = "gzip",
  dbW_digits = 2,
  verbose = FALSE
)
}
\arguments{
\item{dir_data}{directory containing Livneh data}

\item{dir_temp}{the database directory}

\item{site_ids}{the sites to gather weather data for}

\item{coords}{the coordinates for each site in \var{WGS84} format}

\item{start_year}{the start year in the sequence of data to gather}

\item{end_year}{the end year in the sequence of data to gather}

\item{f_check}{flag to check for errors in file structure -
\code{TRUE}
                          for check else no integrity check}

\item{backup}{flag to create a backup of the weather data prior
to insertion in the database
(can create large files) - \code{TRUE} for backup
else no backup}

\item{comp_type}{the compression type of the data to
be inserted into the database}

\item{run_parallel}{whether the extraction should be ran in parallel}

\item{num_cores}{the number of cores to use if parallel}
}
\description{
Extracts daily gridded weather data, including precipitation,
             maximum temperature and minimum temperature from the Livneh
             database: a 1/16 degree gridded weather database that contains
             data for the years 1915 - 2011.
}
\references{
\href{http://www.esrl.noaa.gov/psd/data/gridded/data.livneh.html}{Livneh Weather Website}
}
\author{
Charles Duso    \email{cd622@nau.edu}
}
