% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Spatial_Functions.R
\docType{methods}
\name{extract_rSFSW2}
\alias{extract_rSFSW2}
\alias{extract_rSFSW2,Raster,vector,character-method}
\alias{extract_rSFSW2,Raster,matrix,character-method}
\alias{extract_rSFSW2,Raster,data.frame,character-method}
\alias{extract_rSFSW2,Raster,SpatialPoints,character-method}
\alias{extract_rSFSW2,Raster,Raster,character-method}
\alias{extract_rSFSW2,SpatialPolygons,SpatialPoints,character-method}
\alias{extract_rSFSW2,character,ANY,character-method}
\alias{extract_rSFSW2,SpatialPolygons,SpatialPolygons,character-method}
\alias{extract_rSFSW2,SpatialPolygons,vector,character-method}
\alias{extract_rSFSW2,SpatialPolygons,matrix,character-method}
\title{Extract raster data for point or cell locations}
\usage{
\S4method{extract_rSFSW2}{Raster,vector,character}(x, y, type, ...)

\S4method{extract_rSFSW2}{Raster,matrix,character}(x, y, type, ...)

\S4method{extract_rSFSW2}{Raster,data.frame,character}(x, y, type, ...)

\S4method{extract_rSFSW2}{Raster,SpatialPoints,character}(x, y, type, ...)

\S4method{extract_rSFSW2}{Raster,Raster,character}(x, y, type, ...)

\S4method{extract_rSFSW2}{SpatialPolygons,SpatialPoints,character}(x, y, type,
  ...)

\S4method{extract_rSFSW2}{character,ANY,character}(x, y, type, ...)

\S4method{extract_rSFSW2}{SpatialPolygons,SpatialPolygons,character}(x, y, type,
  ...)

\S4method{extract_rSFSW2}{SpatialPolygons,vector,character}(x, y, type, ...)

\S4method{extract_rSFSW2}{SpatialPolygons,matrix,character}(x, y, type, ...)
}
\arguments{
\item{x}{A \linkS4class{Raster} object from which data are extracted.}

\item{y}{Locations for which data, provided by \code{x}, are extracted.}

\item{type}{A character string. One of 'point' or 'cell'. \itemize{
\item If \code{type == "point"}, then y represents point locations
      by a two-column matrix or data.frame, by \linkS4class{SpatialPoints},
      or by a numeric vector of cell numbers.
\item If \code{type == "cell"}, then y represents cells locations, see
      \code{\link{extract_SFSW2_cells_from_raster}}.
}}
}
\description{
Several methods defined for different use cases described below.
}
\seealso{
\code{\link[raster]{extract}}
}
