% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Functions.R
\name{extract_SFSW2_cells_from_raster}
\alias{extract_SFSW2_cells_from_raster}
\title{Extract the weighted mean (and sample quantiles) for raster cells or rectangles.}
\usage{
extract_SFSW2_cells_from_raster(x, y, ...)
}
\arguments{
\item{x}{A \code{\link[raster:Raster-class]{raster::Raster}} object from which data
are extracted}

\item{y}{Either A \code{\link[raster:RasterLayer-class]{raster::RasterLayer}} OR
raster resolution (of rectangles) as a numeric vector of length two or a matrix
with two columns.
If a \code{\link[raster:RasterLayer-class]{raster::RasterLayer}}, then values of
\code{data} are resampled and extracted for \code{!NA} cells.
If the latter, then the vector or matrix represents the rectangle extent/resolution
in x- and y-coordinates.
If a matrix, then rows must match \code{coords}.}

\item{\dots}{\itemize{
  \item \code{method} A character string. The method argument passed to
    \code{reaggregate_raster}. Default is \var{\dQuote{block}} which is the fastest.
  \item \code{coords} Cell centers (corresponding to \code{!NA} cells of \code{y})
    that are represented by a two-column matrix of \var{xy}-coordinates. If not
    provided, then \code{y} must be a \code{RasterLayer} and cell centers are
    extracted from \code{y}.
  \item \code{probs} A numeric vector of probabilities with values in \code{[0, 1]} at
    which sample quantiles are returned.
}}
}
\value{
A matrix with rows corresponding to the \code{!NA} cells of \code{y} and
  columns to layers of \code{x}.
}
\description{
Extract the weighted mean (and sample quantiles) for raster cells or rectangles.
}
\seealso{
\code{\link[raster]{extract}}
}
