% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{extract_soil_CONUSSOIL}
\alias{extract_soil_CONUSSOIL}
\title{\var{\sQuote{CONUS-SOIL}} is a rasterized and controlled
\var{\sQuote{STATSGO}} dataset; information for 11 soil are layers available.}
\usage{
extract_soil_CONUSSOIL(
  MMC,
  sim_size,
  sim_space,
  project_paths,
  fnames_in,
  resume,
  verbose,
  default_SOM = 0
)
}
\arguments{
\item{default_SOM}{A numeric value. The default value is
0 \var{SOM} [mass proportion].}
}
\description{
\var{\sQuote{CONUS-SOIL}} is a rasterized and controlled
\var{\sQuote{STATSGO}} dataset; information for 11 soil are layers available.
}
\section{Note}{
 It appears that \var{NODATA} values are recorded as 0s.

 We convert what \var{\sQuote{CONUS-SOIL}} calls
  \var{\dQuote{bulk density}} to matric density with equation 20 from
  Saxton et al. 2006: \eqn{bulkd = matricd * (1 - rockvol) + rockvol * 2.65}
  If this variable is indeed \var{\dQuote{bulk density}}, then equation 20
  (Saxton et al. 2006) would give negative values
}

\references{
Miller, D. A. and R. A. White. 1998. A conterminous United States
 multilayer soil characteristics dataset for regional climate and hydrology
 modeling. Earth Interactions 2:1-26.
}
