% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Spatial_Functions.R
\name{extract_swsf_cells_from_raster}
\alias{extract_swsf_cells_from_raster}
\title{Extract the weighted mean (and sample quantiles) for raster cells or rectangles.}
\usage{
extract_swsf_cells_from_raster(x, y, ...)
}
\arguments{
\item{x}{A raster* object from which data are extracted}

\item{y}{Either A RasterLayer OR raster resolution (of rectangles) as a numeric vector
of length two or a matrix with two columns.
If a RasterLayer, then values of \code{data} are resampled and extracted for !NA
cells.
If the latter, then the vector or matrix represents the rectangle extent/resolution
in x- and y-coordinates.
If a matrix, then rows must match \code{coords}.}

\item{\dots}{\itemize{
  \item \code{method} A character string. The method argument passed to
    \code{reaggregate_raster}. Default is 'block' which is the fastest.
  \item \code{coords} Cell centers (corresponding to !NA cells of \code{y}) that are
    represented by a two-column matrix of xy coordinates. If not provided, then
    extracted from \code{y}.
  \item \code{probs} A numeric vector of probabilities with values in \code{[0,1]} at
    which sample quantiles are returned.
}}
}
\value{
A matrix with rows corresponding to the !NA cells of \code{y} and columns to
  layers of \code{x}.
}
\description{
Extract the weighted mean (and sample quantiles) for raster cells or rectangles.
}
\seealso{
\code{\link[raster]{extract}}
}
