% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_ClimateDownscaling.R
\name{ExtractClimateChangeScenarios}
\alias{ExtractClimateChangeScenarios}
\title{Extract climate scenarios}
\usage{
ExtractClimateChangeScenarios(sim_scens, climDB_metas, sim_time,
  climate.ambient, sim_size, SWRunInformation, sw_input_treatments, fnames_in,
  project_paths, opt_parallel, verbose = FALSE, print.debug = FALSE)
}
\description{
Extract climate scenarios
}
