% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_ClimateDownscaling.R
\name{ExtractClimateWizard}
\alias{ExtractClimateWizard}
\title{Extract climate scenarios from downloaded ClimateWizard.org data}
\usage{
ExtractClimateWizard(climDB_metas, SFSW2_prj_meta, SFSW2_prj_inputs,
  verbose = FALSE)
}
\description{
Extract climate scenarios from downloaded ClimateWizard.org data
}
