% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_Elevation.R
\name{ExtractData_Elevation}
\alias{ExtractData_Elevation}
\title{Extract elevation data}
\usage{
ExtractData_Elevation(exinfo, SWRunInformation, sim_size, how_determine_sources,
  sim_space, dir_ex_dem, fnames_in, resume, verbose)
}
\description{
Extract elevation data
}
