% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_MeanMonthlyClimate.R
\name{ExtractData_MeanMonthlyClimate}
\alias{ExtractData_MeanMonthlyClimate}
\title{Extract mean monthly climate data: cloud cover, relative humidity, and wind speed}
\usage{
ExtractData_MeanMonthlyClimate(exinfo, SWRunInformation, sim_size,
  sw_input_cloud_use, sw_input_cloud, how_determine_sources, sim_space,
  project_paths, fnames_in, opt_chunks, resume, verbose, prepd_CFSR, sim_time,
  opt_parallel)
}
\description{
Extract mean monthly climate data: cloud cover, relative humidity, and wind speed
}
