% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_Soils.R
\name{ExtractData_Soils}
\alias{ExtractData_Soils}
\title{Extract soil characteristics}
\usage{
ExtractData_Soils(exinfo, SWRunInformation, sim_size, sw_input_soillayers,
  sw_input_soils_use, sw_input_soils, how_determine_sources, sim_space,
  dir_ex_soil, fnames_in, resume, verbose, opt_parallel)
}
\description{
Extract soil characteristics
}
