% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_WeatherDB.R
\name{ExtractGriddedDailyWeatherFromMaurer2002_NorthAmerica}
\alias{ExtractGriddedDailyWeatherFromMaurer2002_NorthAmerica}
\title{Extract gridded daily weather from Maurer et al. 2002 (updated in 2010) for North
 American sites}
\usage{
ExtractGriddedDailyWeatherFromMaurer2002_NorthAmerica(dir_data, cellname,
  startYear, endYear)
}
\value{
An invisible zero. A list of which each element represents one year of daily
   weather data of class \linkS4class{swWeatherData}. The list is copied to the
   weather database. Units are [degree Celsius] for temperature and [cm / day] and for
   precipitation.
}
\description{
Extract gridded daily weather from Maurer et al. 2002 (updated in 2010) for North
 American sites
}
