% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_WeatherDB.R
\name{ExtractGriddedDailyWeatherFromNRCan_10km_Canada}
\alias{ExtractGriddedDailyWeatherFromNRCan_10km_Canada}
\title{Extract gridded daily weather from NR Canada for Canadian sites}
\usage{
ExtractGriddedDailyWeatherFromNRCan_10km_Canada(dir_data, site_ids,
  coords_WGS84, start_year, end_year, dir_temp = tempdir(),
  dbW_compression_type = "gzip", opt_parallel)
}
\value{
An invisible zero. A list of which each element represents one year of daily
   weather data of class \linkS4class{swWeatherData}. The list is copied to the
   weather database. Units are [degree Celsius] for temperature and [cm / day] and for
   precipitation.
}
\description{
Extract gridded daily weather from NR Canada for Canadian sites
}
