% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mathematical_Functions.R
\name{fun_kLargest}
\alias{fun_kLargest}
\title{Find the k-largest values (and apply a function to these values)}
\usage{
fun_kLargest(x, fun = NULL, k = 10L, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{fun}{A function which requires one argument. \code{fun} will be applied to
the k-largest values of \code{x}.}

\item{k}{An integer value. The k-largest value(s) of \code{x} will be used. The largest
value will be used if 0 or negative.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be stripped
before the computation proceeds.}

\item{\dots}{Optional arguments to be passed to \code{fun}}
}
\value{
A vector with the k-largest values of \code{x} if \code{is.null(fun)},
   otherwise the result of applying \code{fun} to the k-largest values.
}
\description{
Find the k-largest values (and apply a function to these values)
}
