% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{gather_objects_for_export}
\alias{gather_objects_for_export}
\title{Export \var{R} objects to \var{MPI} or socket workers}
\usage{
gather_objects_for_export(varlist, list_envs)
}
\arguments{
\item{varlist}{A vector of R object names to export.}

\item{list_envs}{A list of environments in which to search for the R objects.}
}
\description{
Export \var{R} objects to \var{MPI} or socket workers
}
