% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_OutputDatabase_Functions.R
\name{get.SeveralOverallVariables}
\alias{get.SeveralOverallVariables}
\title{Get data of variables in the overall aggregation table for one of the climCat rows (combining 'Current' and ensembles)}
\usage{
get.SeveralOverallVariables(fdbrSFSW2, fdbrSFSW2ens, climCat, responseName,
  MeanOrSD = "Mean", i_climCat = 1, whereClause = NULL)
}
\description{
Get data of variables in the overall aggregation table for one of the climCat rows (combining 'Current' and ensembles)
}
