% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_OutputDatabase_Functions.R
\name{get.SeveralOverallVariables_Ensemble}
\alias{get.SeveralOverallVariables_Ensemble}
\title{Get data of variables in the overall aggregation table for one of the ensembles}
\usage{
get.SeveralOverallVariables_Ensemble(fdbSWSF, fdbSWSFens, responseName,
  MeanOrSD = "Mean", fam, level, whereClause = NULL)
}
\description{
Get data of variables in the overall aggregation table for one of the ensembles
}
