% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{get.SeveralOverallVariables_Scenario}
\alias{get.SeveralOverallVariables_Scenario}
\title{Get data of variables in the overall aggregation table for one of the
scenarios}
\usage{
get.SeveralOverallVariables_Scenario(fdbrSFSW2, responseName,
  MeanOrSD = "Mean", scenario = "Current", whereClause = NULL)
}
\description{
Get data of variables in the overall aggregation table for one of the
scenarios
}
