% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_OutputDatabase_Functions.R
\name{get.Table}
\alias{get.Table}
\title{Get data-part for an entire table for one of the climCat rows (combining 'Current' and ensembles)}
\usage{
get.Table(fdbSWSF, fdbSWSFens, climCat, responseName, MeanOrSD = "Mean",
  i_climCat = 1, whereClause = NULL, addPid = FALSE)
}
\description{
Get data-part for an entire table for one of the climCat rows (combining 'Current' and ensembles)
}
