% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_OutputDatabase_Functions.R
\name{get.Table_Ensemble}
\alias{get.Table_Ensemble}
\title{Get header and data for an entire table for one of the ensembles}
\usage{
get.Table_Ensemble(fdbSWSF, fdbSWSFens, responseName, MeanOrSD = "Mean", fam,
  level, whereClause = NULL, header = FALSE)
}
\description{
Get header and data for an entire table for one of the ensembles
}
