% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_OutputDatabase_Functions.R
\name{get.Table_Scenario}
\alias{get.Table_Scenario}
\title{Get header and data for an entire table for one of the scenarios}
\usage{
get.Table_Scenario(fdbSWSF, responseName, MeanOrSD = "Mean",
  scenario = "Current", whereClause = NULL, header = FALSE)
}
\description{
Get header and data for an entire table for one of the scenarios
}
