% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorCalculations.R
\name{get_BareSoilEvapCoefs}
\alias{get_BareSoilEvapCoefs}
\title{Calculate bare-soil evaporation coefficients based on soil texture and store
in soil input file}
\usage{
get_BareSoilEvapCoefs(SFSW2_prj_meta, SFSW2_prj_inputs, runIDs_adjust,
  resume = TRUE, verbose = FALSE)
}
\description{
Calculate bare-soil evaporation coefficients based on soil texture and store
in soil input file
}
