% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{get_DailyGCMdata_netCDF}
\alias{get_DailyGCMdata_netCDF}
\title{Extract all daily \var{GCM} projection values for precipitation, minimum and
maximum temperature from each one \var{netCDF} file.}
\usage{
get_DailyGCMdata_netCDF(
  i_tag,
  climDB_meta,
  is_sequential,
  ncFiles,
  startyear,
  endyear,
  lon,
  lat,
  df_site_template,
  var_map
)
}
\value{
A list with three 3-dimensional arrays
\var{\dQuote{tmax}}, \var{\dQuote{tmin}}, and \var{\dQuote{prcp}}.
Units are [degree Celsius] for temperature and [cm / day] for precipitation.
}
\description{
Extract all daily \var{GCM} projection values for precipitation, minimum and
maximum temperature from each one \var{netCDF} file.
}
