% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_OutputDatabase_Functions.R
\name{get_fieldnames}
\alias{get_fieldnames}
\title{Access data from a database}
\usage{
get_fieldnames(responseName, fields.header, fields.iTable)
}
\description{
Access data from a database
}
