% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{get_fnames_dbTempOut}
\alias{get_fnames_dbTempOut}
\title{Locate file names of temporary output database files}
\usage{
get_fnames_dbTempOut(dir_out_temp, ...)
}
\description{
Locate file names of temporary output database files
}
