% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{get_GCMdata_netCDF}
\alias{get_GCMdata_netCDF}
\title{Extract \var{GCM} projection from a \var{netCDF} file}
\usage{
get_GCMdata_netCDF(i_tag, ts_mons, dpm, gcm, scen, rip, lon, lat,
  startyear, endyear, climDB_meta, ...)
}
\value{
A list of one data.frame object with 5 columns and names of
  \var{\dQuote{year}}, \var{\dQuote{month}}, \var{\dQuote{tmax}}, \var{\dQuote{tmin}},
  and \var{\dQuote{prcp}}. Each row is one day.
Units are [degree Celsius] for temperature and [cm / day] and [cm / month],
respectively, for precipitation.
}
\description{
Extract \var{GCM} projection from a \var{netCDF} file
}
