% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_KilledBySoilLayers}
\alias{get_KilledBySoilLayers}
\title{Determine if all conditions across rooted soil layers are deadly}
\usage{
get_KilledBySoilLayers(relevantLayers, kill_conditions)
}
\arguments{
\item{relevantLayers}{An integer vector, usually of length 365 or 366
(days).}

\item{kill.conditions}{A m x p logical matrix with
\code{m >= length(relevantLayers)} and p represents the number of
simulated soil layers, i.e., \code{p >= max(relevantLayers, na.rm = TRUE)}.}
}
\value{
A logical vector of the length of \code{relevantLayers} with
 values containing \code{NA} for days when conditions were not evaluated,
 \code{TRUE} if all relevant soil layers (columns) of \code{kill.conditions}
 were \code{TRUE}, and with \code{FALSE} otherwise
}
\description{
Function that checks whether all relevant (those with roots) soil layers
 are under conditions of mortality (kill.conditions) for each day of a
 given year
}
\details{
\code{relevantLayers} takes either \code{NA} if no soil layers should be
 considered (e.g., because not yet germinated), or an integer number
 between 1 and the number of simulated soil layers. The number indicates
 the depth to which a seedling has grown roots and over which layers
 \code{kill.conditions} will be evaluated.
}
\section{Note}{
 The \pkg{Rcpp} version of the function is about 165x
 faster than the version previous to commit
 \var{6344857a9cdb08acf68fa031c43cf4a596613aad} 'Small speed improvements'
 and about 70x faster than the R version. The \pkg{Rcpp} version also
 reduced the memory footprint by a factor of 200.
}

\examples{
 # The \pkg{Rcpp} function is equivalent to the following R version
    get_KilledBySoilLayers_R <- function(relevantLayers, kill.conditions) {
      vapply(seq_along(relevantLayers), function(k) {
          if (all(is.finite(relevantLayers[k]))) {
            all(as.logical(kill.conditions[k, seq_len(relevantLayers[k])]))
          } else NA
        }, FUN.VALUE = NA)
   }

}
\references{
Schlaepfer, D.R., Lauenroth, W.K. & Bradford, J.B. (2014).
 Modeling regeneration responses of big sagebrush (Artemisia tridentata)
 to abiotic conditions. Ecol Model, 286, 66-77.
}
