% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{get_MonthlyGCMdata_netCDF}
\alias{get_MonthlyGCMdata_netCDF}
\title{Extract monthly \var{GCM} projection from a \var{netCDF} file}
\usage{
get_MonthlyGCMdata_netCDF(
  i_tag,
  time,
  dpm,
  gcm,
  scen,
  rip,
  lon,
  lat,
  startyear,
  endyear,
  climDB_meta,
  ncg,
  nct,
  ncFiles
)
}
\value{
A list of one data.frame object with 5 columns and names of
  \var{\dQuote{year}}, \var{\dQuote{month}},
  \var{\dQuote{tmax}}, \var{\dQuote{tmin}}, and \var{\dQuote{prcp}}.
Each row represents one month.
Units are [degree Celsius] for temperature and [cm / month]
for precipitation.
}
\description{
Extract monthly \var{GCM} projection from a \var{netCDF} file
}
