% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{get_Pid_from_temptxt}
\alias{get_Pid_from_temptxt}
\title{Extract \var{P_id} from content of temporary output files}
\usage{
get_Pid_from_temptxt(str, k = -1, verbose = FALSE)
}
\description{
\var{P_id} values are expected to be at the first position of values, e.g.,
\code{"INSERT INTO \"aggregation_overall_sd\" VALUES (1139776,NULL,..."}
where \code{P_id = 1139776}
}
