% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_SimulationWorld.R
\name{get_simulation_time}
\alias{get_simulation_time}
\title{Determine maximal span of simulation years across all experimental and design
treatments}
\usage{
get_simulation_time(st, SFSW2_prj_inputs)
}
\arguments{
\item{st}{An object as returned from the function
\code{\link{setup_simulation_time}}.}

\item{SFSW2_prj_inputs}{An object as returned from function
\code{\link{process_inputs}}.}
}
\value{
The object \code{st} augmented with two named elements \itemize{
  \item \code{overall_simstartyr} which is the earliest year requested by any
  input \item \code{overall_endyr} which is the last year requested by any
  input }
}
\description{
Determine maximal span of simulation years across all experimental and design
treatments
}
