% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{get_SoilLayerID_from_temptxt}
\alias{get_SoilLayerID_from_temptxt}
\title{Extract soil layer ID from content of temporary output files}
\usage{
get_SoilLayerID_from_temptxt(str, k = -1)
}
\description{
Soil layer ID values are expected to be at the second position of values,
e.g., \code{"INSERT INTO \"aggregation_overall_sd\" VALUES
(1139776,NULL,..."} where \code{sl = NULL}
}
