% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Simulation_Project.R
\name{global_args_simulate_SOILWAT2_experiment}
\alias{global_args_simulate_SOILWAT2_experiment}
\title{List of objects to export which are required by simulate_SOILWAT2_experiment and are
 not in rSWSF (sorted alphabetically)}
\usage{
global_args_simulate_SOILWAT2_experiment()
}
\description{
List of objects to export which are required by simulate_SOILWAT2_experiment and are
 not in rSWSF (sorted alphabetically)
}
