% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indices.R
\name{indices}
\alias{indices}
\alias{it_exp}
\alias{it_site}
\alias{it_Pid}
\alias{it_exp2}
\alias{it_site2}
\alias{it_Pid0}
\alias{it_sim0}
\alias{it_sim2}
\alias{it_scen2}
\title{Index functions}
\usage{
it_exp(isim, runN)

it_exp2(pid, runN, scN)

it_site(isim, runN)

it_site2(pid, runN, scN)

it_Pid(isim, runN, sc, scN)

it_Pid0(iexp, isite, runN, sc, scN)

it_sim0(iexp, isite, runN)

it_sim2(pid, scN)

it_scen2(pid, scN)
}
\arguments{
\item{isim}{An integer vector. A subset of \code{runIDs_todo} which is a
subset of \code{runIDs_total}, i.e., a consecutive index across loop1 and
loop2b}

\item{runN}{An integer value. The number of runs/sites set up in the main
input file, i.e., \eqn{runN == runsN_main}.}

\item{pid}{An integer value. The index value across loop1, loop2b, and loop3
combined, i.e., a unique consecutive identification number across all
simulation runs.}

\item{scN}{An integer value. The number of (climate) scenarios used in the
project, i.e., \eqn{scN == sim_scens[["N"]]}.}

\item{sc}{An integer value. The index along loop3 \code{1...scN}.}

\item{iexp}{An integer value. The index value along loop1 \code{1...expN}.}

\item{isite}{An integer value. The index value along loop2b
\code{1...runsN_sites}.}
}
\value{
The function \code{it_exp} returns an integer value of the index in
  loop 1 based on the position across loops 1+2b; invariant to
  \code{include_YN}.

The function \code{it_site} returns an integer value out of
  \code{runIDs_sites}, i.e., the position in loop 2b based on position
  across loops 1+2b; invariant to \code{include_YN}.

The function \code{it_Pid} returns an integer value of the index
  across all loops 1+2a+3, invariant to \code{include_YN}. A consecutive
  identification number for each possible \pkg{rSFSW2} simulation--used as
  the ID for the output database.
}
\description{
Index functions
}
\section{NOTE}{
 Do not change the indices without adjusting the design of the
  output databases!
}

\section{Simulation runs}{
 \itemize{
 \item Simulations are run over three nested loops \itemize{
   \item loop1 \code{(1...expN)} nested in loop2b \code{(1...runsN_sites)}
     nested in loop3 \code{(1...scN)} \itemize{
     \item Note: loop3 (along scenarios) occurs within the function
       \code{\link{do_OneSite}}
     \item Note: loop2b is a subset of loop2a \code{(1...runsN_main)}}
   \item column \var{\sQuote{include_YN}} reduces \code{site_id} to
     \code{runIDs_sites}
   \item \code{site_id} and \code{P_id} are invariant to
     \var{\sQuote{include_YN}}}

 \item Main input file: column \var{\sQuote{include_YN}} selects rows
   which are included in the simulation \itemize{
   \item Note: rows of the main input file correspond to rows of the
     treatment input file
   \item column \code{site_id} == consecutive identification numbers of all
     rows in the main file; this is treated as a unique (and stable)
     identifier of a site
   \item \code{runsN_main} == number of rows in the main file
   \item \code{runIDs_main} == consecutive identification numbers along
     \code{runsN_main}
   \item \code{runsN_sites} == number of rows in the main file that are
     included; \code{runsN_sites <= max(site_id)} and
     \code{runsN_sites == length(runIDs_sites)}
   \item \code{runIDs_sites} == values of \code{runIDs_main} which are
     included}

 \item Experimental input file: each row defines a condition which is
   applied to every \code{runIDs_sites} \itemize{
   \item \code{expN} == number of experimental treatments}

 \item The function \code{\link{do_OneSite}} will be called n-times with
   \code{n = runsN_call} \itemize{
   \item \code{runsN_job == (number of included sites) x
     (number of experimental treatments) == runsN_sites x expN}
   \item \code{runIDs_job} == consecutive identification numbers along
     \code{runsN_job}
   \item \code{runsN_total == (number of sites) x (number of experimental
     treatments) == runsN_main x expN}
   \item \code{runIDs_total} == consecutive identification numbers along
    \code{runsN_total}
   \item \code{runIDs_done} == values of \code{runIDs_total} that have
     already been processed by \code{\link{do_OneSite}}
   \item \code{runIDs_todo} == values of \code{runIDs_total} that await
     simulation by \code{\link{do_OneSite}}
   \item \code{runsN_todo} == number of \code{runIDs_total} that await
     simulation by \code{\link{do_OneSite}}}

 \item The function \code{\link{do_OneSite}} could be called n-times with
   \code{n = runsN_incl} if all \var{\sQuote{include_YN}} were on \itemize{
   \item \code{runsN_incl} == \code{runsN_total}}

 \item The variable \code{climate.conditions} defines climate conditions
   that are applied to each \code{runIDs_total} \itemize{
   \item \code{scN} == number of climate conditions}

 \item A grand total of \code{n = runsN_Pid} \pkg{rSFSW2} runs could be
   carried out (\var{n} == number of rows in the output database) \itemize{
   \item \code{runsN_Pid == max(P_id) == runsN_total x scN ==
      runsN_main x expN x scN}
   \item \code{P_id} == a unique consecutive identification number for each
     possible \pkg{rSFSW2} simulation; used as the ID for the output
     database}
}
}

