% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Status_Trackers.R
\name{init_intracker}
\alias{init_intracker}
\title{Initialize a new input status tracker object}
\usage{
init_intracker()
}
\value{
A data.frame
}
\description{
Initialize a new input status tracker object
}
