% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Simulation_Project.R
\name{init_rSFSW2_project}
\alias{init_rSFSW2_project}
\title{Initialize a rSFSW2 project (setup description file)}
\usage{
init_rSFSW2_project(SFSW2_prj_meta, fmeta, verbose = TRUE)
}
\arguments{
\item{SFSW2_prj_meta}{A list or environment as generated from a file comparable to
\code{file.path(system.file("demo", package = "rSFSW2"), "SFSW2_project_descriptions.R")}}
}
\value{
An updated version of \code{SFSW2_prj_meta}
}
\description{
Initialize a rSFSW2 project (setup description file)
}
