% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Mathematical_Functions.R
\name{intersect2}
\alias{intersect2}
\title{The intersection on any number of vectors}
\usage{
intersect2(...)
}
\arguments{
\item{...}{Any number of vectors or a list of vectors.}
}
\value{
A vector of the same mode as inputs.
}
\description{
The intersection on any number of vectors
}
\seealso{
\code{\link{intersect}}
}
