% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mathematical_Functions.R
\name{is.natural}
\alias{is.natural}
\title{Test whether input represents a natural number}
\usage{
is.natural(x)
}
\arguments{
\item{x}{An integer, numeric, or complex vector, matrix, or array.}
}
\value{
A logical value.
}
\description{
Test whether input represents a natural number
}
