% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Project.R
\name{is_project_script_file_recent}
\alias{is_project_script_file_recent}
\title{Compare elements and 1-st level structure of project script file with
installed \pkg{rSFSW2}-package version}
\usage{
is_project_script_file_recent(dir_prj,
  script = "SFSW2_project_descriptions.R", ...)
}
\arguments{
\item{dir_prj}{A character string. Path the simulation project folder.}

\item{script}{A character string. Name of the script file to compare.}

\item{...}{Passed to \code{\link{load_project_description}} to pre-fill
new environment.}
}
\value{
A logical value.
}
\description{
Compare elements and 1-st level structure of project script file with
installed \pkg{rSFSW2}-package version
}
\examples{
\dontrun{
is_project_script_file_recent(
  dir_prj = SFSW2_prj_meta[["project_paths"]][["dir_prj"]])}
}
