% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_Soils.R
\name{ISRICWISE12_calc_weightedMeanForSimulationCell}
\alias{ISRICWISE12_calc_weightedMeanForSimulationCell}
\title{Calculate the weighted mean soil variables from ISRIC-WISE v1.2 for one simulation cell}
\usage{
ISRICWISE12_calc_weightedMeanForSimulationCell(i, i_sim_cells_SUIDs, sim_soils,
  layer_N, layer_Nsim, ldepth, dat_wise, nvars)
}
\arguments{
\item{i}{An integer value. The cell number.}

\item{i_sim_cells_SUIDs}{A named numeric vector. A row of of the data.frame returned
by \code{\link{ISRICWISE12_extract_SUIDs}}.}

\item{sim_soils}{A named numeric vector. First element is 'i' and second is 'depth'.
The following elements represent the soil variables (currently, "density", "sand",
"clay", "rock", "carbon"; see \code{\link{prepare_ExtractData_Soils}}) for each of
the maximally possible soil layers (see \code{SFSW2_glovars[["slyrs_maxN"]]}).
Input is an empty template.}

\item{layer_N}{An integer value. The number of soil layers in the ISRIC-WISE v1.2
dataset (i.e., 5, see \code{\link{do_ExtractSoilDataFromISRICWISEv12_Global}}).}

\item{layer_Nsim}{An integer value. The number of soil layers of a \code{rSFSW2} project
representing the ISRIC-WISE v1.2 dataset (i.e., 6, see
\code{\link{do_ExtractSoilDataFromISRICWISEv12_Global}}).}

\item{ldepth}{An integer vector. The depth limits of the extracted \code{rSFSW2}
project soil layers including zero where \code{layer_Nsim + 1 == length(ldepth)}.}

\item{dat_wise}{A data.frame representing the disk file \code{WISEsummaryFile.csv}.}

\item{nvars}{An integer value. The number of soil variables extracted from ISRIC-WISE
v1.2 (currently, 5; see \code{\link{prepare_ExtractData_Soils}}).}
}
\description{
Calculate the weighted mean soil variables from ISRIC-WISE v1.2 for one simulation cell
}
