% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{ISRICWISE_calc_weightedMeanForSimulationCell}
\alias{ISRICWISE_calc_weightedMeanForSimulationCell}
\title{Calculate weighted mean soil variables from one of the ISRIC-WISE data bases for one
   simulation cell}
\usage{
ISRICWISE_calc_weightedMeanForSimulationCell(i, i_sim_cells_SUIDs, sim_soils,
  layer_N, layer_Nsim, ldepth, dat_wise, nvars, var_tags)
}
\arguments{
\item{i}{An integer value. The cell number.}

\item{i_sim_cells_SUIDs}{A named numeric vector. A row of of the data.frame returned
by \code{\link{ISRICWISE_extract_SUIDs}}.}

\item{sim_soils}{A named numeric vector. First element is 'i' and second is 'depth'.
The following elements represent the soil variables (currently, "density", "sand",
"clay", "rock", "carbon"; see \code{\link{prepare_ExtractData_Soils}}) for each of
the maximally possible soil layers (see \code{SFSW2_glovars[["slyrs_maxN"]]}).
Input is an empty template.}

\item{layer_N}{An integer value. The number of soil layers in the dataset (i.e., 5,
see \code{\link{do_ExtractSoilDataFromISRICWISE_Global}}).}

\item{layer_Nsim}{An integer value. The number of soil layers of a \code{rSFSW2} project
representing the dataset (i.e., 6, see
\code{\link{do_ExtractSoilDataFromISRICWISE_Global}}).}

\item{ldepth}{An integer vector. The depth limits of the extracted \code{rSFSW2}
project soil layers including zero where \code{layer_Nsim + 1 == length(ldepth)}.}

\item{dat_wise}{A data.frame representing the disk file \code{WISEsummaryFile.csv}.}

\item{nvars}{An integer value. The number of soil variables extracted from a ISRIC-WISE
dataset (currently, 5; see \code{\link{prepare_ExtractData_Soils}}).}

\item{var_tags}{A vector of character string. The column names for 'suid', bulk
density ('density'), 'sand', 'clay', coarse fragments ('rock'), and organic carbon
content ('carbon').}
}
\description{
Calculate weighted mean soil variables from one of the ISRIC-WISE data bases for one
   simulation cell
}
