% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{ISRICWISE_extract_SUIDs}
\alias{ISRICWISE_extract_SUIDs}
\title{A wrapper for \code{reaggregate_raster} designed to work with the rasters of the
ISRIC-WISE datasets versions 5-arcmin v1.2 and 30-arcsec v1.0}
\usage{
ISRICWISE_extract_SUIDs(i, res = c(0, 0), grid, sp_sites, att = NULL)
}
\arguments{
\item{i}{An integer value. The index to select a location from among \code{sp_sites}
and the corresponding resolution \code{res}.}

\item{res}{A numeric vector of length two or a matrix with two columns. The x- and
y-extent of the rectangle(s) for which to extract values.}

\item{grid}{A \linkS4class{RasterLayer} object with one layer. The raster from which
values are extracted.}

\item{sp_sites}{A \linkS4class{SpatialPoints} object. This object is used to extract
the coordinates of the i-th location.}

\item{att}{A character string. Which variable in the RAT table should be returned.
If \code{NULL} then extracted values of \code{grid} are returned.}
}
\value{
A list with four elements
 \describe{
   \item{i}{An integer value. The location index.}
   \item{SUIDs_N}{An integer vector. The number of unique values within the rectangle
     of \code{x}.}
   \item{SUID}{A numeric vectors. The sorted unique values.}
   \item{fraction}{A numeric vector. The relative areas covered by \code{values}.}
 }
}
\description{
A wrapper for \code{reaggregate_raster} designed to work with the rasters of the
ISRIC-WISE datasets versions 5-arcmin v1.2 and 30-arcsec v1.0
}
\seealso{
\code{\link{reaggregate_raster}}
}
