% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kill_seedling}
\alias{kill_seedling}
\title{Determine seedling survival in the first season (\var{\sQuote{ss1s}})}
\usage{
kill_seedling(ss1s, ry_year_day, ry_useyrs, y, doy)
}
\description{
Determine seedling survival in the first season (\var{\sQuote{ss1s}})
}
\section{Note}{
 The \pkg{Rcpp} version of the function is about 270x faster
 for vectors of length 365 and 12,000x faster for vectors of length 11,000
 than the R version. The \pkg{Rcpp} version also reduced the memory
 footprint by a factor of >> 3080.

 Previous name \code{setFALSE_SeedlingSurvival_1stSeason}.
}

\section{C code}{
 \code{ss1s} is a pointer to the data and the original
 vector will get altered; one would need for a deep copy:
 \code{LogicalVector out = clone(ss1s)}
}

\examples{
 # The \\pkg{Rcpp} function is equivalent to the following R version
   kill_seedling_R <- function(ss1s, ry_year_day, ry_useyrs, y,
     doy) {
     ss1s[ry_year_day == ry_useyrs[y]][doy] <- FALSE
     ss1s
   }

}
\references{
Schlaepfer, D.R., Lauenroth, W.K. & Bradford, J.B. (2014).
 Modeling regeneration responses of big sagebrush (Artemisia tridentata)
 to abiotic conditions. Ecol Model, 286, 66-77.
}
