% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_databases.R
\name{list.dbVariablesOfAllTables}
\alias{list.dbVariablesOfAllTables}
\title{List tables and variables of a database}
\usage{
list.dbVariablesOfAllTables(dbName)
}
\description{
List tables and variables of a database
}
