% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_datafiles.R
\name{load_preprocessed_inputs}
\alias{load_preprocessed_inputs}
\title{Load pre-processed simulation project inputs}
\usage{
load_preprocessed_inputs(SFSW2_prj_meta, verbose = FALSE)
}
\arguments{
\item{SFSW2_prj_meta}{A list.}

\item{verbose}{A logical value.}
}
\value{
A list \code{SFSW2_prj_inputs}.
}
\description{
Load pre-processed simulation project inputs
}
