% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Project.R
\name{load_project_description}
\alias{load_project_description}
\title{Load a project description script}
\usage{
load_project_description(fmetar, ...)
}
\value{
An environment containing the named objects generated by the script
}
\description{
Load a project description script
}
