% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{local_weatherDirName}
\alias{local_weatherDirName}
\title{Get name of weather file from output database}
\usage{
local_weatherDirName(i_sim, runN, scN, dbOutput)
}
\description{
Get name of weather file from output database
}
