% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{make_dbOutput}
\alias{make_dbOutput}
\title{Create \var{\sQuote{dbOutput}} if requested and/or not already present}
\usage{
make_dbOutput(SFSW2_prj_meta, SFSW2_prj_inputs, verbose = FALSE)
}
\value{
An integer value. The number of fields in the
  \var{\sQuote{overall_aggregation}} tables minus 1 (i.e., the field
  \var{\dQuote{P_id}} is not counted here)
}
\description{
Create \var{\sQuote{dbOutput}} if requested and/or not already present
}
\section{NOTE}{
 Do not change the design of the output database without
  adjusting the index functions \code{it_Pid}, \code{it_exp}, and
  \code{it_site} (see part 4)
}

