% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_WeatherDB.R
\name{make_dbW}
\alias{make_dbW}
\title{Create and populate a 'Rsoilwat31' daily weather SQLite database}
\usage{
make_dbW(SWSF_prj_meta, SWRunInformation, opt_parallel, opt_chunks, opt_behave,
  deleteTmpSQLFiles, verbose = FALSE)
}
\description{
Create and populate a 'Rsoilwat31' daily weather SQLite database
}
