% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testproject_Functions.R
\name{make_test_output_reference}
\alias{make_test_output_reference}
\title{Copy output database of a test project to reference folder}
\usage{
make_test_output_reference(dir_test, dir_ref = NULL,
  SFSW2_version = NULL)
}
\arguments{
\item{dir_test}{A character string. Path to test project folder.}

\item{dir_ref}{A character string. Path to folder with reference database.}

\item{SFSW2_version}{A character string. The version ID of the simulation
framework as reported by the file \code{DESCRIPTION}.}
}
\value{
A logical value. \code{TRUE} if successful.
}
\description{
This function is called for its side effect of copying a file to the
reference folder.
}
