% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{move_dbTempOut_to_dbOut}
\alias{move_dbTempOut_to_dbOut}
\title{Moves simulation output that was written to temporary SQL-databases to a final
output SQL-database}
\usage{
move_dbTempOut_to_dbOut(SFSW2_prj_meta, t_job_start, opt_parallel, opt_behave,
  opt_out_run, opt_verbosity, chunk_size = -1L, dir_out_temp = NULL,
  check_if_Pid_present = FALSE)
}
\value{
Invisibly the number of temporary SQL-databases.
}
\description{
Speed tests suggest that the chunking option slows the process down considerably;
thus, the default for \code{chunk_size} turns the chunking off.
}
