% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Project.R
\name{move_output_to_dbOutput}
\alias{move_output_to_dbOutput}
\title{Move temporary output data to output databases}
\usage{
move_output_to_dbOutput(SFSW2_prj_meta, t_job_start, opt_parallel, opt_behave,
  opt_verbosity, dir_out_temp = NULL)
}
\arguments{
\item{dir_out_temp}{A character string. The path to temporary output files.
If \code{NULL}, then temporary output files are assumed to be located at
\code{SFSW2_prj_meta[["project_paths"]][["dir_out_temp"]]}. This cannot be \code{NULL}
unless all simulation runs have finished (to prevent overwriting of temporary output
files of a potential concurrent run).}
}
\description{
Move temporary output data to output databases
}
\section{Notes}{
 'concatenation' may be much faster if temporary text files are not constructed
around SQL insert statements, but instead as data.frames. Text files containing
data.frames may be much faster with checks for duplicate P_id entries and could be
inserted at once (instead of line by line) with the command
  RSQLite::dbWriteTable(con, name = table, value = "path/to/db-file", append = TRUE)

 The variables 'pids_inserted' and 'pids2_inserted' become quickly
  very large and may then be too large for available memory

 It only makes sense to follow `opt_out` option `dbOutCurrent_from_dbOut`
}

