% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{mpi_work}
\alias{mpi_work}
\title{Rmpi work function for calling \code{do_OneSite}}
\usage{
mpi_work(verbose = FALSE)
}
\arguments{
\item{verbose}{A logical value.}
}
\description{
Rmpi work function for calling \code{do_OneSite}
}
\section{Note}{

 If an error occurs, then the worker will likely not report back to master because
 it hangs in miscommunication and remains idle (check activity, e.g., with \code{top}).
}

\references{
based on the example file \href{http://acmmac.acadiau.ca/tl_files/sites/acmmac/resources/examples/task_pull.R.txt}{'task_pull.R' by ACMMaC}
}
