% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Project.R
\name{obtain_existing_rSFSW2_project}
\alias{obtain_existing_rSFSW2_project}
\title{Load description and data of an existing \pkg{rSFSW2} simulation project}
\usage{
obtain_existing_rSFSW2_project(
  path = ".",
  objects = c("SFSW2_prj_meta", "SFSW2_prj_inputs", "SWRunInformation"),
  fmetar = "SFSW2_project_descriptions.R"
)
}
\arguments{
\item{path}{A character string. The path to the simulation project.}

\item{objects}{A vector of character strings. The objects to return.}
}
\value{
A named list with the requested \code{objects}.
}
\description{
This function does not create missing files or change existing files.
}
