% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Project.R
\name{populate_rSFSW2_project_with_data}
\alias{populate_rSFSW2_project_with_data}
\title{Populate \pkg{rSFSW2} project with input data}
\usage{
populate_rSFSW2_project_with_data(
  SFSW2_prj_meta,
  opt_behave,
  opt_parallel,
  opt_chunks,
  opt_out_run,
  opt_verbosity
)
}
\description{
Populate \pkg{rSFSW2} project with input data
}
