% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Miscellaneous_Functions.R
\name{PotentialNaturalVegetation_CompositionShrubsC3C4_Paruelo1996}
\alias{PotentialNaturalVegetation_CompositionShrubsC3C4_Paruelo1996}
\title{Calculate potential natural vegetation composition for the shrub, C3 grass, and C4
 grass components}
\usage{
PotentialNaturalVegetation_CompositionShrubsC3C4_Paruelo1996(MAP_mm, MAT_C,
  monthly.ppt, monthly.temp, dailyC4vars, isNorth, shrub_limit,
  use_Annuals_Fraction, Annuals_Fraction, use_C4_Fraction, C4_Fraction,
  use_C3_Fraction, C3_Fraction, use_Shrubs_Fraction, Shrubs_Fraction,
  use_Forbs_Fraction, Forbs_Fraction, use_BareGround_Fraction,
  BareGround_Fraction)
}
\description{
Calculate potential natural vegetation composition for the shrub, C3 grass, and C4
 grass components
}
