% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Vegetation.R
\name{PotNatVeg_MonthlyBiomassPhenology_from_Climate}
\alias{PotNatVeg_MonthlyBiomassPhenology_from_Climate}
\title{Adjust mean monthly biomass values of grass and shrub functional groups by climate
 relationships}
\usage{
PotNatVeg_MonthlyBiomassPhenology_from_Climate(tr_VegBiom,
  do_adjBiom_by_temp = FALSE, do_adjBiom_by_ppt = FALSE,
  fgrass_c3c4ann = c(1, 0, 0), growing_limit_C = 4, isNorth = TRUE,
  MAP_mm = 450, mean_monthly_temp_C = c(rep(growing_limit_C - 1, 2),
  rep(growing_limit_C + 1, 8), rep(growing_limit_C - 1, 2)))
}
\arguments{
\item{tr_VegBiom}{A data.frame with 12 rows (one for each month) and columns
\code{X.Biomass}, \code{X.Amount.Live}, \code{X.Perc.Live}, and \code{X.Litter} where
\code{X} are for the functional groups shrubs, \code{X = Sh}; C3-grasses,
\code{X = C3}; C4-grasses, \code{X = C4}; and annuals, \code{X = Annual} containing
default input values.}

\item{do_adjBiom_by_temp}{A logical value. If \code{TRUE} then monthly phenology is
adjusted by temperature.}

\item{do_adjBiom_by_ppt}{A logical value. If \code{TRUE} then monthly biomass is
adjusted by precipitation.}

\item{fgrass_c3c4ann}{A numeric vector of length 3. Relative contribution [0-1] of
the C3-grasses, C4-grasses, and annuals functional groups. The sum of
\code{fgrass_c3c4ann} is 1.}

\item{growing_limit_C}{A numeric value. Mean monthly temperatures equal or above this
limit are here considered suitable for growth (growing season). Default value is 4 C.}

\item{isNorth}{A logical value. \code{TRUE} for locations on northern hemisphere.}

\item{MAP_mm}{A numeric value. Mean annual precipitation in millimeter of the location.}

\item{mean_monthly_temp_C}{A numeric vector of length 12. Mean monthly temperature
in Celcius. The default inputs considered March-October as growing season.}
}
\value{
A list with two elements \code{grass}, \code{shrub}. Each element is a matrix
 with 12 rows (one for each month) and columns \code{Biomass}, \code{Amount.Live},
 \code{Perc.Live}, and \code{Litter}.
}
\description{
Adjust mean monthly biomass values of grass and shrub functional groups by climate
 relationships
}
\section{Default inputs}{

 - shrubs (IM_USC00107648_Reynolds; 70% shrubs, 30% C3): biomass was estimated at
   MAP = 450 mm/yr
 - sgs-grassland (GP_SGSLTER; 12% shrubs, 22% C3, and 66% C4): biomass was estimated at
   MAP = 340 mm/yr
}

\references{
Bradford, J.B., Schlaepfer, D.R., Lauenroth, W.K. & Burke, I.C. (2014).
 Shifts in plant functional types have time-dependent and regionally variable impacts
 on dryland ecosystem water balance. J Ecol, 102, 1408-1418.
}
