% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_databases.R
\name{pragmas}
\alias{pragmas}
\alias{PRAGMA_settings1}
\alias{PRAGMA_settings2}
\alias{set_PRAGMAs}
\title{Set some useful \var{\sQuote{PRAGMA}}}
\usage{
PRAGMA_settings1()

PRAGMA_settings2()

set_PRAGMAs(con, settings)
}
\description{
Set some useful \var{\sQuote{PRAGMA}}
}
\references{
\url{http://www.sqlite.org/pragma.html}
}
