% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_prepare_climatedata_files.R
\name{prepare_climatedata_netCDFs}
\alias{prepare_climatedata_netCDFs}
\title{Process downloaded \var{netCDF} files to concatenate if needed otherwise
move to dedicated directory}
\usage{
prepare_climatedata_netCDFs(dir_code, dir_data, dir_duplicates,
  dir_concatables, dir_delete, dir_scrutinize, dir_out,
  climDB_tag = NULL, climDB_meta = NULL)
}
\description{
Process downloaded \var{netCDF} files to concatenate if needed otherwise
move to dedicated directory
}
\examples{
\dontrun{
dir_prj <- file.path("/Volumes", "BookDuo_12TB", "BigData", "GIS", "Data",
  "Weather_Future", "ESGF")
dir_code <- file.path(dir_prj, "Scripts")
dir_data <- file.path(dir_prj, "Downloads")
dir_duplicates <- file.path(dir_prj, "Duplicates")
dir_concatables <- file.path(dir_prj, "Raw_to_concat")
dir_delete <- file.path(dir_prj, "to_delete")
dir_scrutinize <- file.path(dir_prj, "to_scrutinize")
dir_out <- file.path(dir_prj, "..", "ClimateScenarios", "CMIP5",
  "ESGF_Global")
if (all(dir.exists(dir_prj))) {
  prepare_climatedata_netCDFs(dir_code, dir_data, dir_duplicates,
    dir_concatables, dir_delete, dir_scrutinize, dir_out,
    climDB_tag = "CMIP5_ESGF_Global")
}}

}
