% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_datafiles.R
\name{process_inputs}
\alias{process_inputs}
\title{Load and prepare inputs for a \pkg{rSFSW2} simulation project}
\usage{
process_inputs(project_paths, fnames_in, use_preprocin = TRUE,
  verbose = FALSE)
}
\description{
Load and prepare inputs for a \pkg{rSFSW2} simulation project
}
