% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{rbind_2cols_nonoverlapping}
\alias{rbind_2cols_nonoverlapping}
\title{\code{rbind} two objects, but make sure that there is no duplication in two
indicator columns.}
\usage{
rbind_2cols_nonoverlapping(x1, x2)
}
\arguments{
\item{x1}{A two dimensional numeric object; the first two columns (e.g.,
years, months) are matched up against \code{x2}.}

\item{x2}{Same as \code{x1}.}
}
\value{
The object that would result from \code{rbind(x1, x2)} but where
  any duplication, as determined by the first two columns, is resolved, i.e.,
  copy with \code{NAs} is ignored or arithmetic mean across \code{x1} and
  \code{x2}.
}
\description{
\code{rbind} two objects, but make sure that there is no duplication in two
indicator columns.
}
