% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_datafiles.R
\name{read_SOILWAT2_DefaultInputs}
\alias{read_SOILWAT2_DefaultInputs}
\title{Prepare default inputs from \pkg{rSOILWAT2}}
\usage{
read_SOILWAT2_DefaultInputs()
}
\value{
A \code{\link[rSOILWAT2:swInputData-class]{rSOILWAT2::swInputData}} object.
}
\description{
This function loads the data \code{\link[rSOILWAT2]{sw_exampleData}} from
\pkg{rSOILWAT2}; removes all but one soil layer (to prevent carry-over
effects of deeper layers in sites that have a shallower soil profile
simulated by \pkg{rSFSW2}); turns on soil temperature simulations; and
removes any weather data.
}
