% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_datafiles.R
\name{read_SOILWAT2_FileDefaults}
\alias{read_SOILWAT2_FileDefaults}
\title{Read from disk the default input files of SOILWAT2}
\usage{
read_SOILWAT2_FileDefaults(dir_in_sw, swFiles_tag = "file")
}
\description{
Read from disk the default input files of SOILWAT2
}
