% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{read_time_netCDF}
\alias{read_time_netCDF}
\title{Read and interpret time dimension of a netCDF file with CF 1 or larger}
\usage{
read_time_netCDF(filename)
}
\arguments{
\item{filename}{A character string. The name of a netCDF file.}
}
\value{
A list with six elements:
 \describe{
   \item{calendar}{The calendar type, see
     \href{http://cfconventions.org/cf-conventions/v1.6.0/cf-conventions.html#calendar}{CF-conventions}.}
   \item{unit}{The units of the time dimension, see
     \href{http://cfconventions.org/cf-conventions/v1.6.0/cf-conventions.html#time-coordinate}{CF-conventions}.}
   \item{N}{The number of steps along the time dimension.}
   \item{base}{The start date of the time dimension.}
   \item{start}{A numeric vector representing the first date with named elements
     'year' and 'month'.}
   \item{end}{A numeric vector representing the last date with named elements 'year'
     and 'month'.}
 }
}
\description{
Read and interpret time dimension of a netCDF file with CF 1 or larger
}
