% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{remove_from_onexit_expression}
\alias{remove_from_onexit_expression}
\title{Remove one of possibly several expressions recorded by \code{on.exit}}
\usage{
remove_from_onexit_expression(sysonexit, tag)
}
\arguments{
\item{sysonexit}{An expression. The returned value of a call to
\code{sys.on.exit()} from inside the calling function.}

\item{tag}{A character string. An string identifying which of the recorded
expressions should be removed.}
}
\description{
Remove one of possibly several expressions recorded by \code{on.exit}
}
\seealso{
\code{\link{enable_debug_dump}} for examples
}
