% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Spatial_Functions.R
\name{res_to_polygons}
\alias{res_to_polygons}
\title{Convert resolution/rectangles into SpatialPolygons}
\usage{
res_to_polygons(x, y, ...)
}
\description{
Convert resolution/rectangles into SpatialPolygons
}
