% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSWSF-package.R
\docType{package}
\name{rSWSF}
\alias{rSWSF}
\alias{rSWSF-package}
\alias{rSWSF-package}
\title{Framework for SOILWAT2 simulations: creating simulation runs, executing simulations,
 and aggregating outputs}
\description{
SOILWAT2 is forced by \itemize{
   \item daily time traces: \itemize{
           \item rainfall (cm)
           \item minimum air temperature at 2-m height (C)
           \item maximum air temperature at 2-m height (C)
         }
   \item mean monthly values (climate normals): \itemize{
           \item wind speed at 2-m height (miles/h before SOILWAT v24;
                 m/s starting with SOILWAT v24)
           \item relative humidity at 2-m height (%)
           \item cloud cover (%)
         }
   \item mean vegetation
   \item soil characterization
   \item et cetera
 }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Burke-Lauenroth-Lab/SoilWat_R_Wrapper}
  \item Report bugs at \url{https://github.com/Burke-Lauenroth-Lab/SoilWat_R_Wrapper/issues}
}

}
\author{
\strong{Maintainer}: Daniel Schlaepfer \email{daniel.schlaepfer@unibas.ch}

Other contributors:
\itemize{
  \item Caitlin Andrews [contributor]
  \item Ryan Murphy [contributor]
}

}
