% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Run.R
\name{run_simulation_experiment}
\alias{run_simulation_experiment}
\title{Run a \pkg{rSFSW2} simulation experiment}
\usage{
run_simulation_experiment(SFSW2_prj_inputs, MoreArgs, rSW2_options)
}
\description{
Run a \pkg{rSFSW2} simulation experiment
}
