% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testproject_Functions.R
\name{run_test_projects}
\alias{run_test_projects}
\title{Run test projects}
\usage{
run_test_projects(dir_test, dir_tests, dir_prev = NULL,
  which_tests_torun = seq_along(dir_tests), delete_output = FALSE,
  force_delete_output = FALSE, make_new_ref = FALSE)
}
\arguments{
\item{dir_test}{A character string. Path to overall test project folder.}

\item{dir_tests}{A vector of character strings. Paths to individual test projects.}

\item{dir_prev}{A character string. Paths to directory that should be set when function returns.}

\item{which_tests_torun}{An integer vector. Indices of \code{dir_tests} which will be
carried out.}

\item{delete_output}{A logical value. If \code{TRUE} then output will be deleted unless
problems showed up.}

\item{force_delete_output}{A logical value. If \code{TRUE} then output will be deleted
even if problems showed up.}

\item{make_new_ref}{A logical value. If \code{TRUE} then output database will be saved
as new reference.}
}
\value{
A data.frame where each row represents the outcomes of a test project. The
 The columns return elapsed time in seconds, four logical values \code{has_run},
 \code{has_problems}, \code{made_new_refs}, \code{deleted_output}, and one character
 string \code{referenceDB} of the reference database name against which this run of
 the test project was compared.
}
\description{
Run test projects
}
