% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{SeedlingRootingDepth}
\alias{SeedlingRootingDepth}
\title{Function to calculate rooting depth at given age}
\usage{
SeedlingRootingDepth(age, P0, K, r)
}
\description{
Function to calculate rooting depth at given age
}
