% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{select_suitable_CFs}
\alias{select_suitable_CFs}
\title{Subset a list of netCDF CF file names to specific models, scenarios,
and variables}
\usage{
select_suitable_CFs(
  climDB_files,
  climDB_meta,
  getYears,
  model_name,
  scenario_names
)
}
\description{
Subset a list of netCDF CF file names to specific models, scenarios,
and variables
}
