% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomNumberGenerator.R
\name{set_full_RNG}
\alias{set_full_RNG}
\title{Specify all aspects of the R random number generator}
\usage{
set_full_RNG(seed = NULL, kind = "default", normal.kind = "default")
}
\arguments{
\item{seed}{a single value, interpreted as an integer, or \code{NULL}
    (see \sQuote{Details}).}

\item{kind}{character or \code{NULL}.  If \code{kind} is a character
    string, set \R's RNG to the kind desired.  Use \code{"default"} to
    return to the \R default.  See \sQuote{Details} for the
    interpretation of \code{NULL}.}

\item{normal.kind}{character string or \code{NULL}.  If it is a character
    string, set the method of Normal generation.  Use \code{"default"}
    to return to the \R default.  \code{NULL} makes no change.}
}
\description{
Specify all aspects of the R random number generator
}
\seealso{
\code{\link{set.seed}}, \code{\link{RNGkind}}
}
