% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{set_options_warn_error}
\alias{set_options_warn_error}
\title{Setting global 'warn' and 'error' options}
\usage{
set_options_warn_error(
  debug.warn.level = 1L,
  debug.dump.objects = FALSE,
  dir_prj = ".",
  verbose = FALSE
)
}
\arguments{
\item{debug.warn.level}{An integer value. Sets the \code{warn} option.}

\item{debug.dump.objects}{A logical value. Sets the \code{error} option.
See \code{details} section.}

\item{dir_prj}{A character string. The path at which the \var{RData}
file are saved if \code{debug.dump.objects} is turned on.}

\item{verbose}{A logical value.}
}
\value{
A list of length two with elements 'warn' and 'error' containing
 the status of these two global options before resetting them by this
 function.
}
\description{
Setting global 'warn' and 'error' options
}
\section{Details}{
 Accepted values of \code{debug.warn.level} are \itemize{
 \item  warn < 0: warnings are ignored
 \item  warn = 0: warnings are stored until the top-level function returns
 \item  warn = 1: warnings are printed as they occur
 \item  warn = 2: all warnings are turned into errors.
}
 If \code{debug.dump.objects} is \code{TRUE}, then code will on error dump
 objects and frames to files at path \code{dir_prj}, and (if not in
 interactive mode) quit. To view the dumped frames first attach them with
     \code{load(file.path(dir_prj, "last.dump.rda"))}
 and then browse them with
     \code{debugger(`path/to/file/last.dump.rda`)}
}

