% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2_DataAccess.R
\name{set_requested_rSOILWAT2_InputFlags}
\alias{set_requested_rSOILWAT2_InputFlags}
\title{Assign requested values to rSOILWAT2 flags}
\usage{
set_requested_rSOILWAT2_InputFlags(tasks, swIn, tag, use, values, fun,
  reset = TRUE, default = NA)
}
\arguments{
\item{reset}{A logical value. If \code{TRUE}, then reset flags identified by \code{tag}
and turned off as identified by \code{use} to \code{default}. If \code{FALSE}, then
set  flags identified by \code{tag} and turned on as identified by \code{use} to
corresponding elements of \code{values}; other flags are not changed.}
}
\description{
Assign requested values to rSOILWAT2 flags
}
