% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Simulation_Project.R
\name{setup_rSWSF_project_infrastructure}
\alias{setup_rSWSF_project_infrastructure}
\title{Setup infrastructure (skeleton) for a new rSWSF simulation experiment}
\usage{
setup_rSWSF_project_infrastructure(project_paths, opt_verbosity)
}
\description{
Setup infrastructure (skeleton) for a new rSWSF simulation experiment
}
