% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Spatial_Functions.R
\name{setup_spatial_simulation}
\alias{setup_spatial_simulation}
\title{Set-up information for a spatially aware simulation project}
\usage{
setup_spatial_simulation(SFSW2_prj_meta, SFSW2_prj_inputs,
  use_sim_spatial = FALSE)
}
\description{
Set-up information for a spatially aware simulation project
}
