% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{SoilLayer_at_SoilDepth}
\alias{SoilLayer_at_SoilDepth}
\title{Function to convert soil depth to soil layer}
\usage{
SoilLayer_at_SoilDepth(depth_cm, layers_depth)
}
\description{
Function to convert soil depth to soil layer
}
