% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mathematical_Functions.R
\name{stretch_values}
\alias{stretch_values}
\title{Stretch values}
\usage{
stretch_values(x, lambda = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{lambda}{A numeric value. The stretching factor applied to \code{x}.}
}
\value{
A numeric vector of the size of \code{x}.
}
\description{
Values above the mean of \code{x} are made larger and
values below the mean are made smaller - each by
\code{lambda * dist(x, mean(x))}.
}
