% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSOILWAT2_DataAccess.R
\name{swOutput_access}
\alias{swOutput_access}
\alias{get_Response_aggL}
\alias{get_Temp_yr}
\alias{get_Temp_mo}
\alias{get_Temp_dy}
\alias{get_VPD_mo}
\alias{get_PPT_mo}
\alias{get_PPT_dy}
\alias{get_PET_yr}
\alias{get_PET_mo}
\alias{get_AET_yr}
\alias{get_AET_mo}
\alias{get_AET_dy}
\alias{get_SWE_mo}
\alias{get_SWE_dy}
\alias{get_Inf_yr}
\alias{get_Inf_mo}
\alias{get_Inf_dy}
\alias{get_Esurface_yr}
\alias{get_Esurface_dy}
\alias{get_Interception_yr}
\alias{get_DeepDrain_yr}
\alias{get_DeepDrain_mo}
\alias{get_DeepDrain_dy}
\alias{get_RunOnOff_mo}
\alias{get_RunOnOff_yr}
\alias{get_Vegetation_yr}
\alias{get_CO2effects_yr}
\title{rSOILWAT2 data access functions}
\usage{
get_Response_aggL(response, tscale = c("dy", "dyAll", "mo", "moAll", "yr",
  "yrAll"), scaler = 10, FUN, weights = NULL, x, st, st2, topL, bottomL)

get_Temp_yr(x, st)

get_Temp_mo(x, st)

get_Temp_dy(x, st)

get_VPD_mo(sc, temp.mo, xin, st2)

get_PPT_mo(x, st)

get_PPT_dy(x, st)

get_PET_yr(x, st)

get_PET_mo(x, st)

get_AET_yr(x, st)

get_AET_mo(x, st)

get_AET_dy(x, st)

get_SWE_mo(x, st)

get_SWE_dy(x, st)

get_Inf_yr(x, st)

get_Inf_mo(x, st)

get_Inf_dy(x, st)

get_Esurface_yr(x, st)

get_Esurface_dy(x, st)

get_Interception_yr(x, st)

get_DeepDrain_yr(x, st)

get_DeepDrain_mo(x, st)

get_DeepDrain_dy(x, st)

get_RunOnOff_mo(x, st)

get_RunOnOff_yr(x, st)

get_Vegetation_yr(x, st)

get_CO2effects_yr(x, st)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{swOutput}}.}

\item{st}{An object as returned from the function \code{\link{setup_simulation_time}}.}
}
\description{
rSOILWAT2 data access functions
}
