% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Synchronicity.R
\name{synchronicity}
\alias{synchronicity}
\alias{lock}
\alias{unlock}
\alias{lock_init}
\alias{lock_attempt}
\alias{unlock_access}
\alias{lock_access}
\alias{check_lock_content}
\alias{remove_lock}
\alias{lock_init}
\alias{check_lock_content}
\alias{remove_lock}
\alias{unlock_access}
\alias{lock_attempt}
\alias{lock_access}
\title{Synchronicity with backing store locks}
\usage{
lock_init(fname_lock, id)

check_lock_content(lock)

remove_lock(lock)

unlock_access(lock)

lock_attempt(lock)

lock_access(lock, verbose = FALSE)
}
\arguments{
\item{fname_lock}{A character string. Path to locking directory.}

\item{id}{A R object. Identifier used to test unique write access in locking directory.}

\item{lock}{An object of class \code{SWSF_lock}.}

\item{verbose}{A logical value. If \code{TRUE}, then each attempt at obtaining the lock
is printed.}
}
\value{
An object of class \code{SWSF_lock}.

A logical value. \code{TRUE} if lock file contains proper code of the lock.

The return value of unlinking the directory associated with \code{lock}.

The updated \code{lock} object of class \code{SWSF_lock}.

The updated \code{lock} object of class \code{SWSF_lock}.

The updated \code{lock} object of class \code{SWSF_lock}.
}
\description{
Synchronicity with backing store locks

Create and initialize an object of class \code{SWSF_lock}

Check content of a backing store lock

Remove the files associated with a backing store lock

Unlock a backing store lock

Attempt to obtain access of a backing store lock

Access to the backing store lock is attempted until successfully obtained.
}
\section{Note}{
 My attempts at using lock/unlock functions from \pkg{flock} failed
 because unlock appeared to not free resources. My attempts at using such functions
 from \pkg{synchronicity} failed because of segfaults with
 \code{GetResourceName(m@mutexInfoAddr)}; additionally, \pkg{synchronicity} is
 restricted to unix systems. Instead, I use here ideas based on functions
 from \pkg{Rdsm} and Samuel.
}

