% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mathematical_Functions.R
\name{test_sigmaGamma}
\alias{test_sigmaGamma}
\title{Check that data are within range of an approximated gamma distribution}
\usage{
test_sigmaGamma(data, sigmaN = 6)
}
\arguments{
\item{data}{A numeric vector. Daily values of precipitation.}

\item{sigmaN}{An integer value. A multiplier of \code{stats::sd}.}
}
\description{
Check that data are within range of an approximated gamma distribution
}
\references{
Choi, S. C., and R. Wette. 1969. Maximum Likelihood Estimation of the
  Parameters of the Gamma Distribution and Their Bias. Technometrics 11:683-690.

\url{http://en.wikipedia.org/wiki/Gamma_distribution#Maximum_likelihood_estimation}
}
