% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mathematical_Functions.R
\name{test_sigmaNormal}
\alias{test_sigmaNormal}
\title{Check that data are within range of normal distribution}
\usage{
test_sigmaNormal(data, sigmaN = 6)
}
\arguments{
\item{data}{A numeric vector. Daily values of temperature.}

\item{sigmaN}{An integer value. A multiplier of \code{stats::sd}.}
}
\description{
Check that data are within range of normal distribution
}
