% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Vegetation.R
\name{TranspCoeffByVegType}
\alias{TranspCoeffByVegType}
\title{Lookup transpiration coefficients for grasses, shrubs, and trees per soil layer or per soil depth increment of 1 cm per distribution type for each simulation run and copy values to 'datafile.soils'}
\usage{
TranspCoeffByVegType(tr_input_code, tr_input_coeff, soillayer_no, trco_type,
  layers_depth, adjustType = c("positive", "inverse", "allToLast"))
}
\description{
first row of datafile is label for per soil layer 'Layer' or per soil depth increment of 1 cm 'DepthCM'
second row of datafile is source of data
the other rows contain the data for each distribution type = columns
}
\section{Note}{

 cannot write data from sw_input_soils to datafile.soils
}

