% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{trewartha_climate}
\alias{trewartha_climate}
\title{Main climate classification according to Trewartha}
\usage{
trewartha_climate(meanTmonthly_C)
}
\arguments{
\item{meanTmonthly_C}{A numeric vector of length 12. Mean monthly air
temperature in degree Celsius.}
}
\description{
The main climate classifications based on mean monthly temperature, i.e.,
without group B--dry, include \itemize{
  \item Group A (tropical): sum(months >= 18 C) == 12
  \item Group C (subtropical):
    sum(months >= 10 C) >= 8 & sum(months >= 18 C) < 12
  \item Group D (temperate and continental): sum(months >= 10 C) %in% 4:7
  \item Group E (boreal): sum(months >= 10 C) %in% 1:3
  \item Group F (polar): sum(months >= 10 C) == 0
}
}
\references{
Trewartha, G.T. and Lyle, H.H., 1980: An Introduction to Climate.
  McGraw - Hill, 5th Ed. Appendix: Koeppen's Classification of Climates
}
\seealso{
\code{\link[ClimClass]{koeppen_geiger}}
}
