% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{try.ScenarioWeather}
\alias{try.ScenarioWeather}
\title{Make daily weather for a scenario}
\usage{
try.ScenarioWeather(i, clim_source, is_netCDF, is_NEX, climDB_meta,
  climDB_files, reqGCMs, reqRCPsPerGCM, reqDownscalingsPerGCM, climate.ambient,
  locations, dbW_iSiteTable, dbW_iScenarioTable, compression_type, getYears,
  assocYears, sim_time, opt_DS, project_paths, verbose, print.debug)
}
\description{
A wrapper function for \code{calc.ScenarioWeather} with error control.
}
