% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{try_cell_ISRICWISE}
\alias{try_cell_ISRICWISE}
\title{Wrapping \code{\link{calc_cell_ISRICWISE}} in \code{\link[base]{try}} to
handle errors without stopping a \pkg{rSFSW2} run.}
\usage{
try_cell_ISRICWISE(i, sim_cells_SUIDs, template_simulationSoils, layer_N,
  layer_Nsim, ldepth, dat_wise = dat_wise, nvars, var_tags, val_rocks)
}
\description{
Wrapping \code{\link{calc_cell_ISRICWISE}} in \code{\link[base]{try}} to
handle errors without stopping a \pkg{rSFSW2} run.
}
