% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{try_MonthlyScenarioWeather}
\alias{try_MonthlyScenarioWeather}
\title{Make daily weather for a scenario}
\usage{
try_MonthlyScenarioWeather(
  i,
  clim_source,
  use_CF,
  use_NEX,
  climDB_meta,
  climDB_files,
  reqGCMs,
  reqRCPsPerGCM,
  reqDownscalingsPerGCM,
  climate.ambient,
  locations,
  compression_type,
  getYears,
  assocYears,
  sim_time,
  seeds_DS,
  opt_DS,
  project_paths,
  dir_failed,
  fdbWeather,
  resume,
  verbose,
  print.debug
)
}
\description{
A wrapper function for \code{calc_MonthlyScenarioWeather} with error control.
}
