% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Project.R
\name{update_actions}
\alias{update_actions}
\title{Update actions for simulation project}
\usage{
update_actions(SFSW2_prj_meta, actions = NULL, wipe_dbOutput = FALSE)
}
\arguments{
\item{SFSW2_prj_meta}{An environment.}

\item{actions}{A named list of logical elements. See \code{SFSW2_project_code.R}.}

\item{wipe_dbOutput}{A logical value}
}
\value{
A version of \code{SFSW2_prj_meta} with updated values for element
  \code{prj_todos}.
}
\description{
Update actions for simulation project
}
