% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vegetation.R
\name{update_biomass}
\alias{update_biomass}
\title{Replace selected biomass values of a \linkS4class{swProd}-object}
\usage{
update_biomass(fg = c("Grass", "Shrub", "Tree", "Forb"), use, prod_input,
  prod_default)
}
\arguments{
\item{fg}{A character string. One of the functional groups represented by
\code{rSOILWAT2}}

\item{use}{A logical vector.}
}
\description{
Replace selected biomass values of a \linkS4class{swProd}-object
}
