% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{vp0}
\alias{vp0}
\title{Saturation vapor pressure}
\usage{
vp0(T)
}
\arguments{
\item{T}{A numeric vector of temperature(s) (deg C)}
}
\value{
A numeric vector of length \code{T} of saturation vapor pressure (\var{kPa}) at
   temperature T
}
\description{
Saturation vapor pressure
}
\references{
Yoder, R. E., L. O. Odhiambo, and W. C. Wright. 2005. Effects of
   Vapor-Pressure Deficit and Net-Irradiance Calculation Methods on Accuracy of
   Standardized Penman-Monteith Equation in a Humid Climate Journal of Irrigation and
   Drainage Engineering 131:228-237.
}
