% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{vpd}
\alias{vpd}
\title{Vapor pressure deficit}
\usage{
vpd(Tmin, Tmax, RHmean = NULL)
}
\arguments{
\item{Tmin}{A numeric vector of daily minimum temperature(s) (deg C)}

\item{Tmax}{A numeric vector of daily maximum temperature(s) (deg C)}

\item{RHmean}{A numeric vector of daily mean relative humidity (percentage)}
}
\value{
A numeric vector of length \code{T} of vapor pressure deficit
  (\var{kPa})
}
\description{
Vapor pressure deficit
}
\references{
Yoder, R. E., L. O. Odhiambo, and W. C. Wright. 2005. Effects of
  Vapor-Pressure Deficit and Net-Irradiance Calculation Methods on Accuracy
  of Standardized Penman-Monteith Equation in a Humid Climate Journal of
  Irrigation and Drainage Engineering 131:228-237.
}
