% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mathematical_Functions.R
\name{whereNearest}
\alias{whereNearest}
\title{Index of the closest value in the matrix to the passed in value.}
\usage{
whereNearest(val, matrix)
}
\description{
Index of the closest value in the matrix to the passed in value.
}
