% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Functions.R
\name{align_with_target_grid}
\alias{align_with_target_grid}
\title{Aligns \code{grid_from} with \code{grid_to} for certain cells}
\usage{
align_with_target_grid(grid_from, coords, grid_to, crs_to = NULL)
}
\arguments{
\item{grid_from}{A
\code{\link[raster:RasterLayer-class]{raster::RasterLayer}} object.}

\item{coords}{A matrix of x and y coordinates, or a
\code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object indicating
which cells of projected \code{grid_from} will be used.}

\item{grid_to}{A \code{\link[raster:RasterLayer-class]{raster::RasterLayer}}
object.}

\item{crs_to}{A \code{\link[sp:CRS-class]{sp::CRS}} object or \code{NULL} in
which case it will be extracted from \code{grid_to}.}
}
\value{
A list with two elements \describe{
  \item{x}{A \code{\link[raster:RasterLayer-class]{raster::RasterLayer}}
    object. Cells values are \code{NA} or 1 if they contain \code{points}
    of \code{coords}.}
  \item{index}{An integer vector. The cell numbers of \code{x} that
    correspond to \code{coords}.} }
}
\description{
Aligns \code{grid_from} with \code{grid_to} for certain cells
}
