% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{calc.ScenarioWeather}
\alias{calc.ScenarioWeather}
\title{Extract climate scenario data and downscale to daily weather data}
\usage{
calc.ScenarioWeather(i, clim_source, is_netCDF, is_NEX, climDB_meta,
  climDB_files, reqGCMs, reqRCPsPerGCM, reqDownscalingsPerGCM, climate.ambient,
  locations, dbW_iSiteTable, dbW_iScenarioTable, compression_type, getYears,
  assocYears, sim_time, task_seed, opt_DS, project_paths, verbose, print.debug)
}
\description{
Extract climate scenario data and downscale to daily weather data
}
