% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorCalculations.R
\name{calc_BareSoilEvaporationCoefficientsFromSoilTexture}
\alias{calc_BareSoilEvaporationCoefficientsFromSoilTexture}
\title{Calculate potential bare-soil evaporation coefficients}
\usage{
calc_BareSoilEvaporationCoefficientsFromSoilTexture(layers_depth, sand, clay,
  depth_max_bs_evap_cm = 15)
}
\arguments{
\item{layers_depth}{A numeric vector, matrix, or data.frame. Values describe the lower
soil layer depths in units of centimeters.}

\item{sand}{A numeric vector, matrix, or data.frame. Values are sand contents in units
of mass-percentage / 100.}

\item{clay}{A numeric vector, matrix, or data.frame. Values are clay contents in units
of mass-percentage / 100.}

\item{depth_max_bs_evap_cm}{A numeric value. The maximal soil depth in centimeters from
which bare-soil evaporation is potentially drawing moisture.}
}
\value{
A numeric matrix with potential bare-soil evaporation coefficients where rows
 correspond to sites and columns to soil layers.
}
\description{
Soil texture influence based on re-analysis of data from Wythers et al. 1999.
Default of \code{depth_max_bs_evap} = 15 cm from Torres et al. 2010.
}
\section{Notes}{
 Rows of soil input arguments \code{layers_depth}, \code{sand}, and
 \code{clay} correspond to sites and columns to soil layers. If \code{sand} and/or
 \code{clay} are vectors, then they are converted to 1-row matrices.
 If \code{layers_depth} is a vector, then it is converted to a matrix with as many
 sites/rows as \code{sand} and \code{clay} have. That is the code assumes identical
 soil layer depths for each site. All soil input arguments must have a the same number
 of sites and of soil layers, i.e., identical matrix dimensions.
}

\references{
Torres EA, Calera A (2010) Bare soil evaporation under high evaporation
 demand: a proposed modification to the FAO-56 model. Hydrological Sciences Journal-
 Journal Des Sciences Hydrologiques, 55, 303-315.

Wythers KR, Lauenroth WK, Paruelo JM (1999) Bare-Soil Evaporation Under
 Semiarid Field Conditions. Soil Science Society of America Journal, 63, 1341-1349.
}
