% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_PriorCalculations.R
\name{calc_ExtendSoilDatafileToRequestedSoilLayers}
\alias{calc_ExtendSoilDatafileToRequestedSoilLayers}
\title{Add soil layers (by interpolation) if not already present}
\usage{
calc_ExtendSoilDatafileToRequestedSoilLayers(SFSW2_prj_meta, SFSW2_prj_inputs,
  runIDs_adjust, verbose = FALSE)
}
\description{
Add soil layers (by interpolation) if not already present
}
