% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{calc_SiteClimate}
\alias{calc_SiteClimate}
\title{Calculate several climate variables from daily weather}
\usage{
calc_SiteClimate(weatherList, year.start, year.end, do.C4vars = FALSE,
  simTime2 = NULL)
}
\description{
Calculate several climate variables from daily weather
}
