% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{calc_SiteClimate}
\alias{calc_SiteClimate}
\title{Calculate climate variables from daily weather}
\usage{
calc_SiteClimate(weatherList, year.start, year.end, do.C4vars = FALSE,
  simTime2 = NULL)
}
\arguments{
\item{weatherList}{A list. Each element is an object of class
\linkS4class{swWeatherData} containing daily weather data of a specific year.}

\item{year.start}{An integer value. The first year of the range of years for which
climate variables should be calculated.}

\item{year.end}{An integer value. The last year of the range of years for which
climate variables should be calculated.}

\item{do.C4vars}{A logical value. If \code{TRUE} then additional output is returned.}

\item{simTime2}{An object as returned from function
\code{\link{simTiming_ForEachUsedTimeUnit}}. Only needed if \code{isTRUE(do.C4vars)}.}
}
\value{
A list with named elements \itemize{
  \item{\var{\dQuote{meanMonthlyTempC}}} {A numeric vector of length 12.
   Mean monthly mean daily air temperature in degree Celsius.}
  \item{\var{\dQuote{minMonthlyTempC}}} {A numeric vector of length 12.
    Mean monthly minimum daily air temperature in degree Celsius.}
  \item{\var{\dQuote{maxMonthlyTempC}}} {A numeric vector of length 12.
    Mean monthly maximum daily air temperature in degree Celsius.}
  \item{\var{\dQuote{meanMonthlyPPTcm}}} {A numeric vector of length 12.
    Mean monthly precipitation in centimeters.}
  \item{\var{\dQuote{MAP_cm}}} {A numeric value. Mean annual precipitation in
    centimeters.}
  \item{\var{\dQuote{MAT_C}}} {A numeric value. Mean annual air temperature in
    degree Celsius.}
  \item{\var{\dQuote{dailyTempMin}}} {A numeric vector. If \code{isTRUE(do.C4vars)},
    then minimum daily air temperature in degree Celsius for each day of time period
    between \code{year.start} and \code{year.end}. If \code{!isTRUE(do.C4vars)}, then
    \code{NA}.}
  \item{\var{\dQuote{dailyTempMean}}} {A numeric vector. Similar as for
    \code{dailyTempMin} but for mean daily air temperature.}
  \item{\var{\dQuote{dailyC4vars}}} {If \code{isTRUE(do.C4vars)}, then a named
    numeric vector containing the output of \code{\link{sw_dailyC4_TempVar}}, else
    \code{NA}.}
}
}
\description{
Calculate climate variables from daily weather
}
