% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{calcDeltas}
\alias{calcDeltas}
\title{Calculate Deltas, used for downscaling functionality}
\usage{
calcDeltas(obs.hist.monthly, scen.fut.monthly, opt_DS)
}
\description{
Calculate Deltas, used for downscaling functionality
}
