% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISSM.R
\name{calculate_TimeToGerminate_modifiedHardegree2006NLR}
\alias{calculate_TimeToGerminate_modifiedHardegree2006NLR}
\title{Function to estimate time to germinate for each day of a given year and conditions (temperature, top soil SWP)}
\usage{
calculate_TimeToGerminate_modifiedHardegree2006NLR(RYyear,
  Germination_DuringFavorableConditions, LengthDays_FavorableConditions,
  RYyear_ForEachUsedDay, soilTmeanSnow, swp.TopMean, TmeanJan, param)
}
\description{
Function to estimate time to germinate for each day of a given year and conditions (temperature, top soil SWP)
}
