% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_OutputDatabase_Functions.R
\name{check_outputDB_completeness}
\alias{check_outputDB_completeness}
\title{Check whether dbOutput contains a complete set of output/simulation results}
\usage{
check_outputDB_completeness(dbOutput, dbOutput_current = NULL,
  update_workDB = FALSE, do_DBcurrent = FALSE, opt_parallel,
  dir_out = getwd(), verbose = FALSE)
}
\description{
Check whether dbOutput contains a complete set of output/simulation results
}
