% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Simulation_Project.R
\name{check_rSWSF_project_input_data}
\alias{check_rSWSF_project_input_data}
\title{Attempt to check input data of a rSWSF project for consistency}
\usage{
check_rSWSF_project_input_data(SWSF_prj_meta, SWSF_prj_inputs, opt_verbosity)
}
\description{
Attempt to check input data of a rSWSF project for consistency
}
