% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB_Check.R
\name{check_weatherDB}
\alias{check_weatherDB}
\title{Checks data in a weather database}
\usage{
check_weatherDB(dir_prj, fdbWeather, repeats = 2L,
  do_preprocess_tempfiles = TRUE, n_cores = 20L, startyear = 1979,
  endyear = 2010, seed = NA)
}
\arguments{
\item{dir_prj}{A character string. The directory path the rSFSW2 simulation project.}

\item{fdbWeather}{A character string. The file path of weather database.}

\item{repeats}{An integer value. The number of times each weather object is extracted}

\item{do_preprocess_tempfiles}{A logial value. Set to TRUE, for instance, if a
previous run was prematurely aborted.}

\item{seed}{A seed set, \code{NULL}, or \code{NA}. \code{NA} will not affect
the state of the RNG; \code{NULL} will re-initialize the RNG; and all other values
are passed to \code{\link{set.seed}}.}
}
\description{
Checks data in a weather database
}
