% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_parallel.R
\name{clean_SWSF_cluster}
\alias{clean_SWSF_cluster}
\title{Clean the parallel cluster used for a rSFSW2 simulation project}
\usage{
clean_SWSF_cluster(opt_parallel, verbose = FALSE)
}
\description{
Clean the parallel cluster used for a rSFSW2 simulation project
}
