% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timing_Calls.R
\name{compile_overall_timer}
\alias{compile_overall_timer}
\title{Write timing information to the timing file}
\usage{
compile_overall_timer(timerfile2, dir_out, workersN = 0,
  runs.completed = 0, scenario_No = 0, ensembles.completed = 0,
  delta.overall = NA, delta.outputDB = NA, delta.check = NA,
  delta.ensembles = NA)
}
\description{
Write timing information to the timing file
}
