% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Miscellaneous_Functions.R
\name{convert_precipitation}
\alias{convert_precipitation}
\title{Converts precipitation data to values in cm / month}
\usage{
convert_precipitation(x, unit_conv, dpm)
}
\description{
Converts precipitation data to values in cm / month
}
