% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Miscellaneous_Functions.R
\name{convert_temperature}
\alias{convert_temperature}
\title{Converts temperature data to values in degree Celsius}
\usage{
convert_temperature(x, unit_conv)
}
\description{
Converts temperature data to values in degree Celsius
}
