% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Spatial_Functions.R
\name{crs_units}
\alias{crs_units}
\title{Extracts the 'units' argument from a CRS object}
\usage{
crs_units(CRS)
}
\arguments{
\item{CRS}{A Raster*, Spatial*, CRS, or character object with a coordinate reference
system (CRS).}
}
\value{
A character string or \code{NA}.
}
\description{
Extracts the 'units' argument from a CRS object
}
