% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Functions.R
\name{crs_units}
\alias{crs_units}
\title{Extracts the \var{sQuote{units}} argument from a \var{CRS} object}
\usage{
crs_units(CRS)
}
\arguments{
\item{CRS}{A \code{\link[raster:Raster-class]{raster::Raster}},
\code{\link[sp:Spatial-class]{sp::Spatial}}, \code{\link[sp:CRS-class]{sp::CRS}}, or
character object with a coordinate reference system (\var{CRS}).}
}
\value{
A character string or \code{NA}.
}
\description{
Extracts the \var{sQuote{units}} argument from a \var{CRS} object
}
