% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{dbOutput_ListDesignTables}
\alias{dbOutput_ListDesignTables}
\title{List the design tables of \var{\sQuote{dbOutput}}}
\usage{
dbOutput_ListDesignTables()
}
\description{
List the design tables of \var{\sQuote{dbOutput}}
}
