% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{dbOutput_Tables_have_SoilLayers}
\alias{dbOutput_Tables_have_SoilLayers}
\title{List the available output tables of dbOutput which record output of variables per
 soil layer}
\usage{
dbOutput_Tables_have_SoilLayers(tables = NULL, con = NULL, dbname = NULL)
}
\description{
List the available output tables of dbOutput which record output of variables per
 soil layer
}
