% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_databases.R
\name{dbVacuumRollack}
\alias{dbVacuumRollack}
\title{Perform a vacuum operation if there is a rollback journal present}
\usage{
dbVacuumRollack(con, dbname)
}
\description{
Perform a vacuum operation if there is a rollback journal present
}
