% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_check}
\alias{dbWork_check}
\alias{dbWork_check}
\title{Check run status}
\usage{
dbWork_check(path, runIDs)

dbWork_check(path, runIDs)
}
\value{
A data.frame with three columns 'completed', 'failed', and 'inwork'

A data.frame with three columns 'completed', 'failed', and 'inwork'
}
\description{
Check run status

Check run status
}
