% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{dbWork_todos}
\alias{dbWork_todos}
\title{Extract identification numbers of runs of a \pkg{rSFSW2} simulation project which are
 uncompleted and not \var{\dQuote{inwork}}}
\usage{
dbWork_todos(path)
}
\arguments{
\item{path}{A character string. Path to the folder where the database will be created.}
}
\value{
An integer vector of \code{runIDs}.
}
\description{
Extract identification numbers of runs of a \pkg{rSFSW2} simulation project which are
 uncompleted and not \var{\dQuote{inwork}}
}
