% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_Soils.R
\name{do_ExtractSoilDataFromCONUSSOILFromSTATSGO_USA}
\alias{do_ExtractSoilDataFromCONUSSOILFromSTATSGO_USA}
\title{CONUS-SOIL is a rasterized and controlled STATSGO data set; information for 11 soil
are layers available.}
\usage{
do_ExtractSoilDataFromCONUSSOILFromSTATSGO_USA(MMC, sim_size, sim_space,
  dir_ex_soil, fnames_in, resume, verbose)
}
\description{
CONUS-SOIL is a rasterized and controlled STATSGO data set; information for 11 soil
are layers available.
}
\section{Note(drs)}{
 it appears that NODATA values are recorded as 0
}

\references{
Miller, D. A. and R. A. White. 1998. A conterminous United States
 multilayer soil characteristics dataset for regional climate and hydrology modeling.
 Earth Interactions 2:1-26.
}
