% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{do_ExtractSoilDataFromSSURGO}
\alias{do_ExtractSoilDataFromSSURGO}
\title{Download and extract SSURGO data}
\usage{
do_ExtractSoilDataFromSSURGO(MMC, dir_data_SSURGO, fnames_in, verbose,
  SWRunInformation, print.debug, sim_size, sim_space, dir_ex_soil, resume)
}
\arguments{
\item{lat}{Lattitude (float)}

\item{lon}{Longitude (float)}

\item{s}{Size of the bounding box - default is .001}

\item{soil_data}{A named list of length 2:\preformatted{
(1) "spatial": A SpatialPolygonsDataFrame of soil mapunits in the template.
(2) "tabular": A named list of data.frame's with the SSURGO tabular data}}

\item{soil_data}{A named list of length 2:\preformatted{
(1) "spatial": A SpatialPolygonsDataFrame of soil mapunits in the template.
(2) "tabular": A named list of data.frame's with the SSURGO tabular data}}

\item{keys}{dataframe consisting of the mukey, cokey, and chkeys}

\item{formatted_data}{- see return value of extract_and_format_soil_data}
}
\value{
A raster object that serves as a bounding box

list consisting of the mukey, cokey, and chkey(s)

a data frame with component percent, mukey, cokey, and chkey

matrix containing the fields from chorizon, chfrags, and muaggat

matrix containing the fields from chorizon, chfrags, and muaggat
}
\description{
Download the respective SSURGO area for each site, extract
the data into both spatial and tabular files, choose the most prominent keys,
extract the respective data within the keys, and populate the soil layers and soil texture
slots in the MMC variable with that data.

Prints out why a function call failed
It will also enable the extraction of STATSGO once all sites are finished

Create a raster polygon to grab an area from SSURGO via coordinates

Choose the mukey with the largest summed component percent, the cokey with the largest individual component percent (within the mukey),
then grab all chkeys that match the cokey.

Creates a data frame that contains component percent, mukey, cokey, and chkey.

Grabbing data:
   > Mukey, component percent, and cokey are grabbed from component
   > Cokey and chkey are grabbed from chorizon
   > Chkey is grabbed from chfrags.
       (Fragvol is also grabbed, since at least two fields are needed to maintain a data frame)

Joining data:
   > Component is joined with chorizon by cokey
   > The above data frame is joined with chfrags by chkey

Extract the needed data fields from the input, but only if the data fields match the chosen keys.
Some aspects of this functon are not neccessary, as data could be extracted and immediately populated into MMC,
but storing the data in variables makes for better organization and troubleshooting.
}
\details{
\preformatted{The following columns of data will be extracted:
   > chorizon
       > sandtotal.r (percent)
       > claytotal.r (percent)
       > silttotal.r (percent)
       > dbthirdbar.r (g/cm^3)
       > hzdepb.r (cm)
       > hzname (string)
   > chfrags
       > fragvol.r (mm)
   > muaggatt
       > brockdepmin (cm)
   > component
       > comppct.r

To see how keys are chosen, see function choose_keys}

\preformatted{
                        | SSURGO  |     Ours     | Conversion
> chorizon
       > sandtotal.r    | percent |   fraction   | divide by 100
       > claytotal.r    | percent |   fraction   | divide by 100
       > silttotal.r    | percent |   fraction   | divide by 100
       > dbthirdbar.r   | g/cm^3  |   Mg/m^3     | multiply by 10^9
       > hzdepb.r       |   cm    |      cm      | none
       > hzname         | string  |   string     | none
> chfrags
       > fragvol.r      | percent |   fraction   | divide by 100
> muaggatt
       > brockdepmin    |   cm    |      cm      | none}
}
\note{
FedData cannot grab a single pair of coordinates

Only tabular is needed, but spatial is usually bundled with tabular

Units retreived from SSURGO Metadata:
      http://www.nrcs.usda.gov/wps/PA_NRCSConsumption/download?cid=stelprdb1241114&ext=pdf
}
\author{
Zachary Kramer, \email{kramer.zachary.nau@gmail.com}
}
