% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{downscale.delta}
\alias{downscale.delta}
\title{Downscale with the 'delta approach'}
\usage{
downscale.delta(obs.hist.daily, obs.hist.monthly, scen.hist.monthly,
  scen.fut.monthly, itime, years = NULL, sim_time = NULL,
  opt_DS = list(ppt_type = "detailed", sigmaN = 6, PPTratioCutoff = 10),
  dailyPPTceiling, do_checks = TRUE, ...)
}
\arguments{
\item{obs.hist.daily}{A list. Each element corresponds to one year of
\code{simstartyr:endyr} is an object of
\code{\link[rSOILWAT2:swWeatherData-class]{rSOILWAT2::swWeatherData}}.}

\item{obs.hist.monthly}{A numeric matrix. Monthly time-series of observed weather
calculated from \code{obs.hist.daily} for the years \code{simstartyr:endyr}.}

\item{scen.hist.monthly}{A numeric matrix. Monthly time-series of scenario weather
during the historic time period \code{DScur_startyr:DScur_endyr}}

\item{scen.fut.monthly}{A numeric matrix. Monthly time-series of scenario weather
during the projected time period \code{DSfut_startyr:DSfut_endyr}}

\item{opt_DS}{A named list.}

\item{do_checks}{A logical value. If \code{TRUE} perform several sanity checks on the
data.}
}
\description{
Downscale with the 'delta approach'
}
\references{
Hay, L. E., R. L. Wilby, and G. H. Leavesley. 2000. A comparison of delta
 change and downscaled GCM scenarios for three mountainous basins in the United States.
 Journal of the American Water Resources Association 36:387-397.

Hamlet, A. F., E. P. Salathe, and P. Carrasco. 2010. Statistical
 downscaling techniques for global climate model simulations of temperature and
 precipitation with application to water resources planning studies. Chapter 4. Final
 Report for the Columbia Basin Climate Change Scenarios Project. Climate Impacts Group,
 Center for Science in the Earth System, Joint Institute for the Study of the
 Atmosphere and Ocean, University of Washington, Seattle, WA.
}
