% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_WeatherDB.R
\name{dw_determine_sources}
\alias{dw_determine_sources}
\title{Determine sources of daily weather}
\usage{
dw_determine_sources(dw_source, exinfo, dw_avail_sources, create_treatments,
  sim_size, SWRunInformation, sw_input_treatments_use, sw_input_treatments,
  sw_input_experimentals_use, sw_input_experimentals, sim_time, fnames_in,
  project_paths, verbose = FALSE)
}
\description{
Determine order of priorities (highest priority comes last): i.e., the last entry is
the one that will be used
}
