% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{exit_SFSW2_cluster}
\alias{exit_SFSW2_cluster}
\title{Clean up and terminate the parallel cluster used for a rSFSW2 simulation project}
\usage{
exit_SFSW2_cluster(opt_parallel, verbose = FALSE)
}
\description{
Clean up and terminate the parallel cluster used for a rSFSW2 simulation project
}
