% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{export_objects_to_workers}
\alias{export_objects_to_workers}
\title{Export objects to workers}
\usage{
export_objects_to_workers(obj_env, parallel_backend = c("mpi", "cluster"),
  cl = NULL)
}
\arguments{
\item{parallel_backend}{A character vector, either 'mpi' or 'cluster'}

\item{cl}{A parallel (socket) cluster object}
}
\value{
A logical value. \code{TRUE} if every object was exported successfully.
}
\description{
Export objects to workers
}
