% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Spatial_Functions.R
\name{extract_SFSW2_points_from_shp}
\alias{extract_SFSW2_points_from_shp}
\title{Extract spatial polygon data for point locations}
\usage{
extract_SFSW2_points_from_shp(x, y, fields = NULL, code = NULL, ...)
}
\arguments{
\item{x}{An object inheriting from \linkS4class{SpatialPolygons} from which data are
extracted.}

\item{y}{graphics::points represented by an object inheriting from
\linkS4class{SpatialPoints}.}

\item{fields}{A character vector. If not \code{NULL}, then \code{fields} selects
columns of the extracted object.}

\item{code}{A vector. If not \code{NULL}, then the extracted data are treated as
integer codes of a factor whose levels are encoded by \code{code}.}

\item{file_shp}{A character string. The filename of the shapefile.}
}
\value{
A vector or matrix with length/rows corresponding to the elements of \code{y}
 and available columns requested by \code{fields}. If \code{!is.null(code)},
 then the encoded 'factor levels' are returned.
}
\description{
Extract spatial polygon data for point locations
}
\seealso{
\code{\link[sp]{over}}
}
