% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{extract_soil_ISRIC250m}
\alias{extract_soil_ISRIC250m}
\title{Extracts soils data (clay, sand, \var{bd}, gravel, depth, ...) from \var{.tif} files in
\var{dir_ex_soil} to a list}
\usage{
extract_soil_ISRIC250m(MMC, sim_size, sim_space, dir_ex_soil, fnames_in, resume,
  verbose, default_TOC_GperKG = 0)
}
\arguments{
\item{sim_size}{A list containing information on the runs.}

\item{sim_space}{A list containing information on site locations and
extraction type, i.e. cell or point.}

\item{fnames_in}{A list containing output file locations.}

\item{resume}{Logical whether or not to resume the project if it failed
partially through before.}

\item{verbose}{Logical whether to see additional messages or not as the
function executes.}

\item{default_TOC_GperKG}{A numeric value. The default value is
0 g \var{TOC} per kg soil.}

\item{\var{MMC}}{A list containing the structure for extracted data to be
extracted to}

\item{\var{dir_ex_soil}}{String for the location of the soil files to be
extracted.}
}
\value{
\var{MMC} A list containing the extracted data.
}
\description{
Extracts soils data (clay, sand, \var{bd}, gravel, depth, ...) from \var{.tif} files in
\var{dir_ex_soil} to a list
}
\section{Note}{
 \enumerate{
\item All data is taken at a resolution of 250m, downloadable at
\var{ftp://ftp.soilgrids.org/data/recent/}
   \item Data Extracted Using Hengl. T. et al, 2017: \itemize{
     \item Sand
     \item Clay
     \item Bulk Density
     \item Gravel
   }
   \item Data Extracted Using Shangguan, W. et al, 2016: \itemize{
     \item Depth to Bedrock
   }
   \item Filenames expected to be in format 
   \var{soiltype_M_sllayernumber_250m.tif}
   Examples: \itemize{
      \item Layer 2 clay = \var{clay_M_sl2_250m.tif}
      \item Layer 4 bulk density = \var{bd_M_sl4_250m.tif}
      }
 }
}

\references{
Hengl. T. et al, 2017. SoilGrids250m: Global gridded soil
 information based on machine learning.

Shangguan, W. et al, 2016. Mapping the global depth to bedrock
for land surface modeling.
}
\author{
Nathan Payton- McCauslin. July 2018.
}
