% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_ClimateDownscaling.R
\name{ExtractClimateChangeScenarios}
\alias{ExtractClimateChangeScenarios}
\title{Extract climate scenarios}
\usage{
ExtractClimateChangeScenarios(climDB_metas, SWSF_prj_meta, SWSF_prj_inputs,
  opt_parallel, verbose = FALSE, print.debug = FALSE)
}
\description{
Extract climate scenarios
}
