% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{ExtractClimateChangeScenarios}
\alias{ExtractClimateChangeScenarios}
\title{Extract climate scenarios}
\usage{
ExtractClimateChangeScenarios(climDB_metas, SFSW2_prj_meta, SFSW2_prj_inputs,
  todos, opt_parallel, opt_chunks, resume, verbose = FALSE,
  print.debug = FALSE)
}
\description{
Extract climate scenarios
}
