% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{ExtractClimateChangeScenarios}
\alias{ExtractClimateChangeScenarios}
\title{Extract climate scenarios}
\usage{
ExtractClimateChangeScenarios(
  climDB_metas,
  SFSW2_prj_meta,
  SFSW2_prj_inputs,
  todos,
  opt_parallel,
  opt_chunks,
  verbose = FALSE,
  print.debug = FALSE
)
}
\arguments{
\item{todos}{A logical vector of length \code{runsN_main}. Element locations with
\code{TRUE} indicate to extract climate data for said 'run'. The \code{TRUE} elements
should be a subset of the \code{TRUE}s of \code{SFSW2_prj_inputs[["include_YN"]]}.}
}
\description{
Extract climate scenarios
}
