% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_ClimateDownscaling.R
\name{ExtractClimateWizard}
\alias{ExtractClimateWizard}
\title{Extract climate scenarios from downloaded ClimateWizard.org data}
\usage{
ExtractClimateWizard(sim_scens, SWRunInformation, fnames_in,
  sw_input_climscen_use, sw_input_climscen, sw_input_climscen_values_use,
  sw_input_climscen_values, dir_ex_fut, sim_size, verbose = FALSE)
}
\description{
Extract climate scenarios from downloaded ClimateWizard.org data
}
