% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Elevation.R
\name{ExtractData_Elevation}
\alias{ExtractData_Elevation}
\title{Extract elevation data}
\usage{
ExtractData_Elevation(exinfo, SFSW2_prj_meta, SFSW2_prj_inputs,
  resume = FALSE, verbose = FALSE)
}
\description{
Extract elevation data
}
