% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_Soils.R
\name{ExtractData_Soils}
\alias{ExtractData_Soils}
\title{Extract soil characteristics}
\usage{
ExtractData_Soils(exinfo, SFSW2_prj_meta, SFSW2_prj_inputs, opt_parallel,
  resume, verbose = FALSE)
}
\description{
Extract soil characteristics
}
