% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_WeatherDB.R
\name{ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_swWeather}
\alias{ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_swWeather}
\alias{ExtractDayMet}
\alias{ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_dbW}
\title{Extract gridded daily weather from DayMet for North American sites}
\usage{
ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_swWeather(dir_data, site_ids,
  coords_WGS84, start_year, end_year, dbW_digits)

ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_dbW(dir_data, site_ids,
  coords_WGS84, start_year, end_year, dir_temp = tempdir(),
  dbW_compression_type = "gzip", dbW_digits, verbose = FALSE)
}
\value{
A list of class \linkS4class{swWeatherData} objects.

An invisible zero. A list of which each element represents one year of daily
   weather data of class \linkS4class{swWeatherData}. The list is copied to the
   weather database. Units are [degree Celsius] for temperature and [cm / day] and for
   precipitation.
}
\description{
Extract gridded daily weather from DayMet for North American sites
}
\references{
\href{https://daymet.ornl.gov/}{daymet website}
 publication: Thornton, P.E., Running, S.W., White, M.A. 1997. Generating surfaces of
   daily meteorological variables over large regions of complex terrain. Journal of
   Hydrology 190: 214 - 251. http://dx.doi.org/10.1016/S0022-1694(96)03128-9
 dataset v3: Thornton, P.E., M.M. Thornton, B.W. Mayer, Y. Wei, R. Devarakonda, R.S.
   Vose, and R.B. Cook. 2016. Daymet: Daily Surface Weather Data on a 1-km Grid for
   North America, Version 3. ORNL DAAC, Oak Ridge, Tennessee, USA. Accessed Month DD,
   YYYY. Time period: YYYY-MM-DD to YYYY-MM-DD. Spatial Range: N=DD.DD, S=DD.DD,
   E=DDD.DD, W=DDD.DD. http://dx.doi.org/10.3334/ORNLDAAC/1328
 \href{https://github.com/khufkens/daymetr}{daymetr package}
}
