% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB.R
\name{ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_swWeather}
\alias{ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_swWeather}
\alias{ExtractDayMet}
\alias{ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_dbW}
\title{Extract gridded daily weather from \var{\dQuote{DayMet}} for North American
sites}
\usage{
ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_swWeather(
  dir_data,
  site_ids,
  coords_WGS84,
  start_year,
  end_year,
  dbW_digits
)

ExtractGriddedDailyWeatherFromDayMet_NorthAmerica_dbW(
  dir_data,
  site_ids,
  site_ids_by_dbW,
  coords_WGS84,
  start_year,
  end_year,
  dir_temp = tempdir(),
  dbW_compression_type = "gzip",
  dbW_digits = 4L,
  verbose = FALSE
)
}
\value{
A list of class
  \code{\link[rSOILWAT2:swWeatherData-class]{rSOILWAT2::swWeatherData}}
  objects.

An invisible zero. A list of which each element represents one year
  of daily weather data of class
  \code{\link[rSOILWAT2:swWeatherData-class]{rSOILWAT2::swWeatherData}}. The
  list is copied to the weather database. Units are [degree Celsius] for
  temperature and [cm / day] and for precipitation.
}
\description{
Data are available for years 1980-(last full calendar year) at a 1-km
resolution.
}
\references{
\url{https://daymet.ornl.gov/}

Thornton, P.E., Rupesh Shrestha, M.M. Thornton, S.C. Kao, Y. Wei,
  and B.E. Wilson.
  Developments in a daily gridded meteorological data set for North America
  -- Daymet Version 4. Draft Manuscript

Thornton, P.E., Running, S.W., White, M.A. 1997.
  Generating surfaces of daily meteorological variables over large regions of
  complex terrain. Journal of Hydrology 190: 214 - 251.
  \doi{10.1016/S0022-1694(96)03128-9}

Daymet: Daily Surface Weather Data on a 1-km Grid for North America,
  Version 4. \doi{10.3334/ORNLDAAC/1840}

\url{https://github.com/khufkens/daymetr}
}
