% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{fix_PPTdata_length}
\alias{fix_PPTdata_length}
\title{Adds/removes elements of data}
\usage{
fix_PPTdata_length(data, targetLength)
}
\arguments{
\item{data}{A numeric vector. Daily values of precipitation.}

\item{targetLength}{An integer value.}
}
\value{
A copy of \code{data} with adjusted length.
}
\description{
Adds/removes elements of data until \code{identical(length(data), targetLength)}
}
\details{
Removal of elements is by random sampling. Addition of elements is by randomly
placing them within the sequence of days of \code{data} and them randomly
assigning values of the original \code{data} with replacement.
}
