% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mathematical_Functions.R
\name{fun_kLargest}
\alias{fun_kLargest}
\title{Find the \code{k}-largest/smallest values (and apply a function to these
values)}
\usage{
fun_kLargest(x, largest = TRUE, fun = NULL, k = 10L, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{largest}{A logical value. See return value.}

\item{fun}{A function which requires one argument or \code{"index"}.
\code{fun} will be applied to the \code{k}-largest/smallest values of
\code{x}.}

\item{k}{An integer value. The \code{k}-largest/smallest value(s) of \code{x}
will be used. The largest/smallest value will be used if 0 or negative.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}

\item{\dots}{Optional arguments to be passed to \code{fun}}
}
\value{
A vector of length \code{k}, \itemize{
  \item if \code{is.null(fun)}, then a vector with the \code{k}-largest
    (if \code{largest = TRUE}) or \code{k}-smallest
    (if \code{largest = FALSE}) values of \code{x};
  \item if \code{fun = "index"}, then a vector with indices of the
    \code{k}-largest/smallest values (NOTE: this is truncated to the
    \code{k}-first indices!). } Otherwise, the result of applying \code{fun}
    to the \code{k}-largest/smallest values.
}
\description{
Find the \code{k}-largest/smallest values (and apply a function to these
values)
}
