% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{get.SeveralOverallVariables_Ensemble}
\alias{get.SeveralOverallVariables_Ensemble}
\title{Get data of variables in the overall aggregation table for one of the ensembles}
\usage{
get.SeveralOverallVariables_Ensemble(fdbrSFSW2, fdbrSFSW2ens, responseName,
  MeanOrSD = "Mean", fam, level, whereClause = NULL)
}
\description{
Get data of variables in the overall aggregation table for one of the ensembles
}
