% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB.R
\name{get_DayMet_cellID}
\alias{get_DayMet_cellID}
\title{Obtain information on \var{DayMet} \var{gridcells}}
\usage{
get_DayMet_cellID(x, crs = 4326)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \code{\link[terra:SpatVector-class]{terra::SpatVector}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
\code{\link[raster:Raster-class]{raster::Raster}},
\code{\link[sp:Spatial-class]{sp::Spatial}},
\code{\link[sp:CRS-class]{sp::CRS}},
or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}
}
\value{
A named list with elements: \describe{
  \item{cellID}{
    A character vector identifying \var{daymet}
    \var{gridcell}/pixels with \var{LCC} coordinates of their
    lower-left corner.
  }
  \item{dm_LCC}{
    A numeric matrix with \var{x} and \var{y} coordinates
    of the \var{gridcell} centers in \var{LCC} projection.
  }
  \item{dm_WGS84}{
    A numeric matrix with \var{longitude} and \var{latitude}
    coordinates of the \var{gridcell} centers.
  }
}
}
\description{
Obtain information on \var{DayMet} \var{gridcells}
}
