% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{get_GCMdata_NEX}
\alias{get_GCMdata_NEX}
\title{Download downscaled GCM projections from NEX}
\usage{
get_GCMdata_NEX(i_tag, ts_mons, dpm, gcm, scen, lon, lat, startyear, endyear,
  climDB_meta, ...)
}
\value{
A list of one data.frame object with 5 columns and names of
"year", "month", "tmax", "tmin", and "prcp". Each row is one day.
Units are [degree Celsius] for temperature and [cm / day] and [cm / month], respectively, for precipitation.
}
\description{
Download downscaled GCM projections from NEX
}
