% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB.R
\name{GriddedDailyWeatherFromNCEPCFSR_Global}
\alias{GriddedDailyWeatherFromNCEPCFSR_Global}
\title{Extract gridded daily weather from NCEP/CFSR for sites globally}
\usage{
GriddedDailyWeatherFromNCEPCFSR_Global(site_ids, dat_sites, tag_WeatherFolder,
  start_year, end_year, meta_cfsr, n_site_per_core = 100, opt_parallel,
  rm_temp = TRUE, resume = FALSE, dir_temp = tempdir(),
  dbW_compression_type = "gzip", dbW_digits)
}
\value{
An invisible zero. A list of which each element represents one year of daily
   weather data of class \linkS4class{swWeatherData}. The list is copied to the
   weather database. Units are [degree Celsius] for temperature and [cm / day] and for
   precipitation.
}
\description{
Extract gridded daily weather from NCEP/CFSR for sites globally
}
\section{Daily data}{
 (http://rda.ucar.edu/datasets/ds093.1/):
 ds093.1 NCEP Climate Forecast System Reanalysis (CFSR) Selected Hourly Time-Series
 Products, January 1979 to December 2010, 0.313-deg: 6-hourly.
 \describe{
   \item{maximum temperature}{2 m above ground (Kelvin): 'tmax.gdas.yyyymm.grb2'
       --> max of 4 values per day}
   \item{minimum temperature}{2 m above ground (Kelvin): 'tmin.gdas.yyyymm.grb2'
       --> max of 4 values per day}
   \item{precipitation rate}{ground or water surface (kg m-2 s-1):
       'prate.gdas.yyyymm.grb2' --> sum of 4 values per day which are converted to
       cm / 6-hour}
 }
}

\references{
\href{http://rda.ucar.edu/datasets/ds093.1/}{NCEP/CFSR website}
 publication: Saha, S., et al. 2010. NCEP Climate Forecast System Reanalysis (CFSR)
   Selected Hourly Time-Series Products, January 1979 to December 2010. Research
   Data Archive at the National Center for Atmospheric Research, Computational and
   Information Systems Laboratory. http://dx.doi.org/10.5065/D6513W89.
}
