% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB.R
\name{gridMET_download_and_check}
\alias{gridMET_download_and_check}
\title{Prepare script to download all or missing files of the \var{gridMET} dataset}
\usage{
gridMET_download_and_check(dir_data, desc = gridMET_metadata())
}
\arguments{
\item{dir_data}{A character string. Path to where the \var{gridMET} dataset
is/will be stored on disk.}

\item{desc}{A named list. Describing the \var{gridMET} dataset.}
}
\value{
If all files are available, then a message is printed to the
 R console with that information. Otherwise, the message points to a
 \var{.sh} script that was created at \code{dir_data} which must be run
 separately to download the missing files.
}
\description{
Prepare script to download all or missing files of the \var{gridMET} dataset
}
\section{Notes}{
 The download scripts use \var{wget}, i.e., it must be
  available on your system to work. The scripts are based on the dataset
  repository setup at \url{http://www.climatologylab.org/gridmet.html} as of
  Nov 2019. This dataset has also been know as \var{METDATA}.
}

\examples{
if (exists("SFSW2_prj_meta")) {
  gridMET_download_and_check(
   dir_data = SFSW2_prj_meta[["project_paths"]][["dir_gridMET"]]
 )
}

}
\references{
Abatzoglou, J. T. (2013) Development of gridded surface
  meteorological data for ecological applications and modelling.
  \var{Int. J. Climatol.}, 33: 121–131.
}
