% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indices.R
\name{indices}
\alias{indices}
\alias{it_exp}
\alias{it_site}
\alias{it_Pid}
\alias{it_exp}
\alias{it_exp2}
\alias{it_site}
\alias{it_site2}
\alias{it_Pid}
\alias{it_Pid0}
\alias{it_sim0}
\alias{it_sim2}
\alias{it_scen2}
\title{Index functions}
\usage{
it_exp(isim, runN)

it_exp2(pid, runN, scN)

it_site(isim, runN)

it_site2(pid, runN, scN)

it_Pid(isim, runN, sc, scN)

it_Pid0(iexp, isite, runN, sc, scN)

it_sim0(iexp, isite, runN)

it_sim2(pid, scN)

it_scen2(pid, scN)
}
\arguments{
\item{isim}{An integer vector. A subset of \code{runIDs_todo} which is a subset of
\code{runIDs_total}, i.e., a consecutive index across loop1 and loop2b}

\item{runN}{An integer value. The number of runs/sites set up in the master input file,
i.e., \eqn{runN == runsN_master}.}

\item{pid}{An integer value. The index value across loop1, loop2b, and loop3 combined,
i.e., a unique consecutive identification number across all simulation runs.}

\item{scN}{An integer value. The number of (climate) scenarios used in the project,
i.e., \eqn{scN == sim_scens[["N"]]}.}

\item{sc}{An integer value. The index along loop3 \code{1...scN}.}

\item{iexp}{An integer value. The index value along loop1 \code{1...expN}.}

\item{isite}{An integer value. The index value along loop2b \code{1...runsN_sites}.}
}
\value{
The function \code{it_exp} returns an integer value of the index in loop 1
 based on the position across loops 1+2b; invariant to \code{include_YN}.

The function \code{it_site} returns an integer value out of 'runIDs_sites',
  i.e., the position in loop 2b based on position across loops 1+2b; invariant to
  \code{include_YN}.

The function \code{it_Pid} returns an integer value of the index across all
 loops 1+2a+3, invariant to \code{include_YN}. A consecutive identification number
 for each possible rSFSW2 simulation--used as the ID for the output database.
}
\description{
Index functions
}
\section{NOTE}{

 Do not change the indices without adjusting the design of the output databases!
}

\section{Simulation runs}{
 \itemize{
 \item Simulations are run over three nested loops \itemize{
   \item loop1 (1...expN) nested in loop2b (1...runsN_sites) nested in loop3 (1...scN)
       - Note: loop3 (along scenarios) occurs within the function 'do_OneSite'
       - Note: loop2b is a subset of loop2a (1...runsN_master)
   \item column 'include_YN' reduces 'site_id' to 'runIDs_sites'
   \item 'site_id' and 'P_id' are invariant to 'include_YN'}

 \item Master input file: column 'include_YN' selects rows which are included in the
   simulation \itemize{
   \item Note: rows of the master input file correspond to rows of the treatment input file
   \item column 'site_id' == consecutive identification numbers of all rows in the master
     file; this is treated as a unique (and stable) identifier of a site
   \item runsN_master == number of rows in the master file
   \item runIDs_master == consecutive identification numbers along runsN_master
   \item runsN_sites == number of rows in the master file that are included;
     runsN_sites <= max(site_id) and runsN_sites == length(runIDs_sites)
   \item runIDs_sites == values of runIDs_master which are included}

 \item Experimental input file: each row defines a condition which is applied to every
   runIDs_sites \itemize{
   \item expN == number of experimental treatments}

 \item The function 'do_OneSite' will be called n-times with n = runsN_call \itemize{
   \item runsN_job == (number of included sites) x (number of experimental treatments)
     == runsN_sites x expN
   \item runIDs_job == consecutive identification numbers along runsN_job
   \item runsN_total == (number of sites) x (number of experimental treatments)
     == runsN_master x expN
   \item runIDs_total == consecutive identification numbers along runsN_total
   \item runIDs_done == values of runIDs_total that have already been processed by
     'do_OneSite'
   \item runIDs_todo == values of runIDs_total that await simulation by 'do_OneSite'
   \item runsN_todo == number of runIDs_total that await simulation by 'do_OneSite'}

 \item The function 'do_OneSite' could be called n-times with n = runsN_incl if all
   'include_YN' were on \itemize{
   \item runsN_incl == runsN_total}

 \item The variable 'climate.conditions' defines climate conditions that are applied to
   each 'runIDs_total' \itemize{
   \item scN == number of climate conditions}

 \item A grand total of n = runsN_Pid rSFSW2 runs could be carried out (n == number of
   rows in the output database) \itemize{
   \item runsN_Pid == max(P_id) == runsN_total x scN == runsN_master x expN x scN
   \item P_id == a unique consecutive identification number for each possible rSFSW2
     simulation; used as the ID for the output database}
}
}

