% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Simulation_Project.R
\name{init_rSWSF_project}
\alias{init_rSWSF_project}
\title{Initialize a rSWSF project (setup description file)}
\usage{
init_rSWSF_project(SWSF_prj_meta, fmeta, verbose = TRUE)
}
\arguments{
\item{SWSF_prj_meta}{A list or environment as generated from a file comparable to
\code{file.path(system.file("demo", package = "rSWSF"), "SWSF_project_descriptions.R")}}
}
\value{
An updated version of \code{SWSF_prj_meta}
}
\description{
Initialize a rSWSF project (setup description file)
}
