% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_OutputDatabase_Functions.R
\name{make_dbOutput}
\alias{make_dbOutput}
\title{Create dbOutput if requested and/or not already present}
\usage{
make_dbOutput(dbOutput, prj_todos, opt_agg, SWRunInformation,
  Index_RunInformation, sim_size, create_treatments, create_experimentals,
  sw_input_treatments, sw_input_treatments_use, sw_input_experimentals,
  sim_scens, sim_time, verbose = FALSE)
}
\value{
An integer value. The number of fields in the 'overall_aggregation' tables
  minus 1 (i.e., 'P_id' is not counted here)
}
\description{
Create dbOutput if requested and/or not already present
}
\section{NOTE}{
 Do not change the design of the output database without adjusting the index
  functions 'it_Pid', 'it_exp', and 'it_site' (see part 4)
}

