% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB.R
\name{make_dbW}
\alias{make_dbW}
\title{Create and populate a \pkg{rSOILWAT2} daily weather \var{SQLite} database}
\usage{
make_dbW(
  fdbWeather,
  SWRunInformation,
  runIDs_sites,
  runIDs_sites_by_dbW = NULL,
  ambient_scenario = "Current",
  dbW_compression_type = "gzip",
  verbose = FALSE
)
}
\arguments{
\item{fdbWeather}{A character string. The path to the weather database.}
}
\description{
Create and populate a \pkg{rSOILWAT2} daily weather \var{SQLite} database
}
