% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB.R
\name{make_dbW}
\alias{make_dbW}
\title{Create and populate a \pkg{rSOILWAT2} daily weather \var{SQLite} database}
\usage{
make_dbW(
  SFSW2_prj_meta,
  SWRunInformation,
  opt_parallel,
  opt_chunks,
  resume = TRUE,
  deleteTmpSQLFiles = TRUE,
  verbose = FALSE,
  print.debug = FALSE
)
}
\description{
Create and populate a \pkg{rSOILWAT2} daily weather \var{SQLite} database
}
