% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_WeatherDB.R
\name{make_dbW}
\alias{make_dbW}
\title{Create and populate a 'Rsoilwat31' daily weather SQLite database}
\usage{
make_dbW(fdbWeather, sim_size, SWRunInformation, sim_time, sim_scens,
  project_paths, opt_chunks, resume, deleteTmpSQLFiles, dbW_compression_type,
  tag_WeatherFolder, opt_parallel, prepd_CFSR = NULL, verbose = FALSE)
}
\description{
Create and populate a 'Rsoilwat31' daily weather SQLite database
}
