% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputDatabase.R
\name{missing_Pids_outputDB}
\alias{missing_Pids_outputDB}
\title{Identify P_id for which output is not completely available in the dbOutput}
\usage{
missing_Pids_outputDB(Table, dbname)
}
\description{
Identify P_id for which output is not completely available in the dbOutput
}
