% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{mpi_last}
\alias{mpi_last}
\title{Properly end mpi slaves before quitting R (e.g., at a crash)}
\usage{
mpi_last()
}
\description{
Properly end mpi slaves before quitting R (e.g., at a crash)
}
\section{Notes}{
 code is based on http://acmmac.acadiau.ca/tl_files/sites/acmmac/resources/examples/task_pull.R.txt
}

