% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{mpi_last}
\alias{mpi_last}
\title{Properly end \var{\dQuote{mpi}} workers before quitting R (e.g., at a crash)}
\usage{
mpi_last(gv)
}
\arguments{
\item{gv}{A list with at least one named element \code{p_has}. \code{p_has} is a
logical value and indicates whether call is from a parallel run.}
}
\description{
Properly end \var{\dQuote{mpi}} workers before quitting R (e.g., at a crash)
}
\section{Notes}{
 Code is based on
 \url{http://acmmac.acadiau.ca/tl_files/sites/acmmac/resources/examples/task_pull.R.txt}.
}

\section{Details}{
 \code{gv} will
 usually be the package-level global variable environment \code{SFSW2_glovars}. This
 is because this function is registered as finalizer to the object
 \code{SFSW2_glovars}.
}

