% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{mpi_work}
\alias{mpi_work}
\title{\pkg{Rmpi} work function for calling \code{do_OneSite}}
\usage{
mpi_work(verbose = FALSE)
}
\arguments{
\item{verbose}{A logical value.}
}
\description{
\pkg{Rmpi} work function for calling \code{do_OneSite}
}
\section{Notes}{
 If an error occurs, then the worker will likely not report
  back to main because it hangs in miscommunication and remains idle
  (check activity, e.g., with \code{top}).
}

\section{Details}{

Message tags sent by this function from workers to main: \itemize{
 \item 1 = worker is ready for main to send a task
 \item 2 = worker is done with a task
 \item 3 = worker is exiting
 \item 4 = worker failed with task
}
Message tags from main which this function can receive and understand:
\itemize{
 \item 1 = this communication is a new task
 \item 2 = tells worker to shut down because all tasks are completed
}
}

\references{
based on the example file
  \url{http://acmmac.acadiau.ca/tl_files/sites/acmmac/resources/examples/task_pull.R.txt}
}
