% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractData_ClimateDownscaling.R
\name{PrepareClimateScenarios}
\alias{PrepareClimateScenarios}
\title{Extracts climate change scenarios and downscales monthly to daily time series}
\usage{
PrepareClimateScenarios(SFSW2_prj_meta, SFSW2_prj_inputs, opt_parallel,
  opt_verbosity)
}
\description{
Extracts climate change scenarios and downscales monthly to daily time series
}
