% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkDatabase.R
\name{recreate_dbWork}
\alias{recreate_dbWork}
\title{Re-create dbWork based on dbOutput}
\usage{
recreate_dbWork(path, dbOutput)
}
\arguments{
\item{dbOutput}{A character string. Full name to the output database.}
}
\value{
A logical vector indicating success.
}
\description{
Re-create dbWork based on dbOutput
}
