% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Simulation_Run.R
\name{run_simulation_experiment}
\alias{run_simulation_experiment}
\title{Run a rSFSW2 simulation experiment}
\usage{
run_simulation_experiment(sim_size, SFSW2_prj_inputs, MoreArgs)
}
\description{
Run a rSFSW2 simulation experiment
}
