% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomNumberGenerator.R
\name{set_RNG_stream}
\alias{set_RNG_stream}
\title{Set the seed of a random number generator (stream)}
\usage{
set_RNG_stream(seed = NA)
}
\arguments{
\item{seed}{A vector appropriate for \code{\link{.Random.seed}} of the current RNG; a
single integer or NULL that will be passed to set.seed(); or NA which will not
affect the random number generator.}
}
\description{
This function is to be called by a (parallel) worker using as argument a pre-prepared
 seed from the function \code{\link{generate_RNG_streams}}. Note: it will also set
 RNGkind accordingly to the first element of seed.
}
\seealso{
\code{\link{set.seed}}, \code{\link{RNGkind}}
}
