% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_dbWork.R
\name{setup_dbWork}
\alias{setup_dbWork}
\title{Setup or connect to SQLite-database \code{dbWork} to manage runs fo a SWSF simulation
 project}
\usage{
setup_dbWork(path, sim_size, include_YN, resume = FALSE)
}
\arguments{
\item{resume}{A logical value. If \code{TRUE} and \code{dbWork} exists,
then function connects to the existing database. If \code{FALSE}, then a new database
is created (possibly overwriting an existing one).}
}
\value{
A logical value indicating success/failure of setting up/connecting to
 \code{dbWork} and initializing with \code{runIDs}.
}
\description{
Setup or connect to SQLite-database \code{dbWork} to manage runs fo a SWSF simulation
 project
}
