% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_parallel.R
\name{setup_SWSF_cluster}
\alias{setup_SWSF_cluster}
\title{Set-up a parallel cluster to be used for a rSWSF simulation project}
\usage{
setup_SWSF_cluster(opt_parallel, opt_verbosity, dir_out)
}
\description{
Set-up a parallel cluster to be used for a rSWSF simulation project
}
