% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_SimulationWorld.R
\name{simTiming_ForEachUsedTimeUnit}
\alias{simTiming_ForEachUsedTimeUnit}
\title{Calculate indices along different time steps for simulation time}
\usage{
simTiming_ForEachUsedTimeUnit(st, sim_tscales = c("daily", "weekly",
  "monthly", "yearly"), use_doy_range = FALSE, doy_ranges = list(),
  latitude = 90, account_NorthSouth = TRUE)
}
\arguments{
\item{st}{An object as returned from the function \code{\link{setup_simulation_time}}.}

\item{sim_tscales}{A vector of character strings. One or multiple of \code{c("daily",
"weekly", "monthly", "yearly")}.}

\item{use_doy_range}{A logical value. If \code{TRUE}, then the result is
additional daily indices indicating whether the \var{DOY} is within the days indicated
in the \code{doy_ranges}.}

\item{doy_ranges}{A named list. Aggregation output variables and the daily
\code{c(min, max)} of days you wish to calculate the aggregation over.}

\item{latitude}{A numeric value. The latitude in decimal degrees for which a
hemispheric adjustment is requested; however, the code extracts only the sign.}

\item{account_NorthSouth}{A logical value. If \code{TRUE}, then the result is
corrected for locations on the southern vs. northern hemisphere.}
}
\value{
A named list.
}
\description{
Calculate indices along different time steps for simulation time
}
