% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Simulation_Project.R
\name{simulate_SOILWAT2_experiment}
\alias{simulate_SOILWAT2_experiment}
\title{Carry out a rSFSW2 simulation experiment}
\usage{
simulate_SOILWAT2_experiment(actions, SFSW2_prj_meta, SFSW2_prj_inputs,
  t_job_start, opt_behave, opt_parallel, opt_chunks, opt_out_run, opt_verbosity)
}
\description{
Carry out a rSFSW2 simulation experiment
}
