% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Project.R
\name{simulate_SOILWAT2_experiment}
\alias{simulate_SOILWAT2_experiment}
\title{Carry out a \pkg{rSFSW2} simulation experiment}
\usage{
simulate_SOILWAT2_experiment(
  SFSW2_prj_meta,
  SFSW2_prj_inputs,
  opt_behave,
  opt_parallel,
  opt_chunks,
  opt_out_run,
  opt_verbosity,
  check_dbWork = TRUE
)
}
\description{
Carry out a \pkg{rSFSW2} simulation experiment
}
