% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_Mathematical_Functions.R
\name{stretch_values}
\alias{stretch_values}
\title{Stretch the values}
\usage{
stretch_values(x, lambda = 0)
}
\description{
Values above the mean of \code{x} are made larger and
values below the mean are made smaller - each by \code{lambda * dist(x, mean(x))}.
}
