% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWSF_ExtractData_ClimateDownscaling.R
\name{tryToGet_ClimDB}
\alias{tryToGet_ClimDB}
\title{Organizes the calls (in parallel) which obtain specified scenario weather for the weather database from one of the available GCM sources}
\usage{
tryToGet_ClimDB(is_ToDo, clim_source, is_netCDF, is_NEX, climDB_meta,
  climDB_files, reqGCMs, reqRCPsPerGCM, reqDownscalingsPerGCM, locations,
  getYears, assocYears, project_paths, fdbWeather, opt_parallel,
  climate.ambient, dbW_iSiteTable, dbW_iScenarioTable, dbW_compression_type,
  sim_time, opt_DS, verbose, print.debug)
}
\description{
This function assumes that a whole bunch of global variables exist and contain appropriate values.
}
