% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous_Functions.R
\name{update_intracker}
\alias{update_intracker}
\title{Update input tracker status}
\usage{
update_intracker(ist, tracker, prepared = NULL, checked = NULL,
  clean_subsequent = FALSE)
}
\arguments{
\item{ist}{A data.frame representing an input tracker as generated by function
\code{init_intracker}.}

\item{tracker}{A character string. One of the rownames of \code{ist}.}

\item{prepared}{A logical value or \code{NULL}. If not \code{NULL}, then the requested
tracker will be updated with this value and its time stamp set.}

\item{checked}{A logical value or \code{NULL}. If not \code{NULL}, then the requested
tracker will be updated with this value and its time stamp set.}

\item{clean_subsequent}{A logical value. If \code{TRUE} then trackers in rows greater
than \code{tracker} will be reset to \code{FALSE}.}
}
\value{
The updated data.frame \code{ist}.
}
\description{
Update input tracker status
}
