% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeatherDB.R
\name{update_runIDs_sites_by_dbW}
\alias{update_runIDs_sites_by_dbW}
\title{Lookup IDs of sites as found in a weather database}
\usage{
update_runIDs_sites_by_dbW(
  fdbWeather,
  label_WeatherData_runIDs_sites,
  runIDs_sites_by_dbW = NULL
)
}
\arguments{
\item{fdbWeather}{A character string. The path to the weather database.}

\item{label_WeatherData_runIDs_sites}{A vector of character strings.
The names of the weather data corresponding to each \code{runIDs_sites}.}

\item{runIDs_sites_by_dbW}{A numeric vector. \code{NULL} or previously
identified identification numbers of sites as used by the weather database.}
}
\value{
A numeric vector with identification numbers of sites as used
  by the weather database (replacing argument \code{runIDs_sites_by_dbW}).
  \code{runIDs_sites_by_dbW} is returned unchanged
  if no update was needed or if no connection to the weather database
  can be established (possibly because it does not exist).
}
\description{
Lookup IDs of sites as found in a weather database
}
