% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgraders.R
\name{upgrade_datafile}
\alias{upgrade_datafile}
\alias{upgrade_prodin_v10_to_v11}
\alias{upgrade_soilsin_v11_to_v12}
\title{Upgrade a 'data input file'}
\usage{
upgrade_datafile(file, new_name, insert_after_tag, inserted_colnames)

upgrade_prodin_v10_to_v11(file, new_name = NULL)

upgrade_soilsin_v11_to_v12(file, new_name = NULL)
}
\arguments{
\item{file}{A character string. The file path to the datafile to upgrade.}

\item{new_name}{A character string. The basename of the upgraded file. If \code{NULL}
then the code attempts to replace 'vOLD' with 'vNEW' in \code{file}.}

\item{insert_after_tag}{A character string. A string identifying existing column(s)
after which new columns are added.}

\item{inserted_colnames}{A character vector. Column name(s) to be added to \code{file}.}
}
\value{
A logical value. \code{TRUE} if upgrade was successful.
}
\description{
Add (a) new column(s)

Add a column of critical soil water potential (SWPcrit) for each functional type

Add a column of total organic content (TOC) for each soil layer
}
