% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Project.R
\name{write_scen_description}
\alias{write_scen_description}
\title{Write scenario description table to file on disk}
\usage{
write_scen_description(SFSW2_prj_meta)
}
\description{
Write scenario description table to file on disk
}
